Creates an object of class 'MeteorologyInterpolationData'

Initializes an object for meteorology interpolation over landscapes using weather station data and the methods described in Thornton et al. (1997) and Thornton & Running (1999).

MeteorologyInterpolationData(
  points,
  elevation = NULL,
  slope = NULL,
  aspect = NULL,
  MinTemperature = NULL,
  MaxTemperature = NULL,
  Precipitation = NULL,
  RelativeHumidity = NULL,
  Radiation = NULL,
  WindSpeed = NULL,
  WindDirection = NULL,
  WindFields = NULL,
  params = defaultInterpolationParams()
)

Arguments

points

An object of class SpatialPointsMeteorology, an object of SpatialPointsTopography or an object of class SpatialPoints (see 'Details'').

elevation

A numeric vector with elevation values of weather stations (in meters).

slope

A numeric vector with slope values of weather stations (in degrees). Needed for cross-validation of interpolation routines.

aspect

A numeric vector with aspect values of weather stations (in degrees from North). Needed for cross-validation of interpolation routines.

MinTemperature

A matrix with minimum temperature recordings (in degrees Celsius) for all weather stations (in rows) and all days (in columns).

MaxTemperature

A matrix with maximum temperature recordings (in degrees Celsius) for all weather stations (in rows) and all days (in columns).

Precipitation

A matrix with precipitation recordings (in mm of water) for all weather stations (in rows) and all days (in columns).

RelativeHumidity

A matrix with (mean) relative humidity recordings (in percent) for all weather stations (in rows) and all days (in columns).

Radiation

A matrix with relative radiation recordings (in MJ/m2) for all weather stations (in rows) and all days (in columns). Needed for cross-validation only.

WindSpeed

A matrix with wind speed recordings (in m/s) for all weather stations (in rows) and all days (in columns).

WindDirection

A matrix with wind direction recordings (in degrees from North) for all weather stations (in rows) and all days (in columns).

WindFields

Object of class "list". See function readWindNinjaWindFields.

params

A list containing interpolation parameters.

Value

An object of class MeteorologyInterpolationData.

Details

There are three ways of building an object of MeteorologyInterpolationData:

The first way is using an object of SpatialPointsMeteorology containing both the coordinates and meteorological series of stations. In this case elevation has to be provided, but aspect and slope may be omitted. Parameters MinTemperature to WindDirection can be left as NULL.

The second way is using an object of class of SpatialPointsTopography containing the coordinates of stations and topographic variables. In this case parameters MinTemperature and MaxTemperature will need to be supplied, each being a matrix with weather stations in rows and days in columns, but Precipitation, RelativeHumidity, Radiation, WindSpeed and WindDirection may be left as NULL.

The third way is using an object of SpatialPoints containing the coordinates of stations only. In this case elevation has to be provided, but aspect and slope may be omitted. As in the second case, parameters MinTemperature and MaxTemperature will need to be supplied, each being a matrix with weather stations in rows and days in columns, but other variables may be left as NULL.

References

Thornton, P.E., Running, S.W., 1999. An improved algorithm for estimating incident daily solar radiation from measurements of temperature, humidity, and precipitation. Agric. For. Meteorol. 93, 211–228. doi:10.1016/S0168-1923(98)00126-9.

Thornton, P.E., Running, S.W., White, M. a., 1997. Generating surfaces of daily meteorological variables over large regions of complex terrain. J. Hydrol. 190, 214–251. doi:10.1016/S0022-1694(96)03128-9.

See also

Author

Miquel De Cáceres Ainsa, CREAF