Creates an object of class 'MeteorologyUncorrectedData'

Initializes an object for statistical correction of meteorological data over landscapes.

MeteorologyUncorrectedData(points, reference_data, projection_data, dates,
                           params = defaultCorrectionParams())

Arguments

points
An object of class SpatialPoints.

reference_data
Reference (historic) meteorological data used to calibrate correction factors when compared with observations. A vector of data frames (one per point) or a single data frame containing the meta data (columns dir and filename) of meteorological files that will be read from the disk. Alternatively, a NetCDF file name where points should be read.

projection_data
Projected meteorological data to be corrected. A vector of data frames (one per point) or a single data frame containing the meta data (columns dir and filename) of meteorological files that will be read from the disk. Alternatively, a NetCDF file name where points should be read.

dates
Object of class "Date" describing the time period for which meteorological correction is possible (corresponding to projection_data).

params
A "list" containing correction parameters.

Details

See correction details in vignettes or in correctionpoints.

Value

An object of class MeteorologyUncorrectedData.

Author

Miquel De Cáceres Ainsa, CREAF

See also