Class "SpatialPointsTopography"

An S4 class that represents topography over a grid of coordinates.

Objects from the Class

Objects can be created by calls of the form new("SpatialPointsTopography", ...), or by calls to the function SpatialPointsTopography.

Slots

data:

Object of class "data.frame" containing the elevation (in m), slope (in degrees) and aspect (in degrees from North) of every point.

coords:

Object of class "matrix" with the spatial coordinates.

bbox:

Object of class "matrix" with the boundary box.

proj4string:

Object of class "CRS" with the projection string.

Extends

Class "SpatialPointsDataFrame", directly. Class "SpatialPoints", by class "SpatialPointsDataFrame", distance 2. Class "Spatial", by class "SpatialPointsDataFrame", distance 3.

Methods

[

signature(x = "SpatialPointsTopography", i = "ANY", j = "ANY", ..., drop = TRUE): subsets the spatial points and associated topography; only rows (points) can be subsetted.

Author

Miquel De Cáceres Ainsa, CREAF

See also

Examples

#Structure of the S4 object
showClass("SpatialPointsTopography")
#> Class "SpatialPointsTopography" [package "meteoland"]
#> 
#> Slots:
#>                                                                   
#> Name:         data  coords.nrs      coords        bbox proj4string
#> Class:  data.frame     numeric      matrix      matrix         CRS
#> 
#> Extends: 
#> Class "SpatialPointsDataFrame", directly
#> Class "SpatialPoints", by class "SpatialPointsDataFrame", distance 2
#> Class "Spatial", by class "SpatialPointsDataFrame", distance 3
#> Class "SpatialVector", by class "SpatialPointsDataFrame", distance 3