Add topography data to meteo object

Add topography data to meteo object

add_topo(meteo, topo, verbose = getOption("meteoland_verbosity", TRUE))

Arguments

meteo

meteo object

topo

topo object

verbose

Logical indicating if the function must show messages and info. Default value checks "meteoland_verbosity" option and if not set, defaults to TRUE. It can be turned off for the function with FALSE, or session wide with options(meteoland_verbosity = FALSE)

Value

meteo with the topography info added

Details

When using meteo data without topography info to create an inteprolator, topography must be addded

See also

Examples

# example meteo
data(meteoland_meteo_no_topo_example)
# example topo
data(meteoland_topo_example)
# add topo
with_meteo(meteoland_meteo_no_topo_example) |>
  add_topo(meteoland_topo_example)
#> ℹ Checking meteorology object...
#> ✔ meteorology object ok
#> ℹ Checking topography object...
#> ✔ topography object ok
#> ℹ Adding topography to meteo (by station ID)...
#> ✔ Topography added
#> Simple feature collection with 5652 features and 16 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 0.30565 ymin: 40.55786 xmax: 3.18165 ymax: 42.77011
#> Geodetic CRS:  WGS 84
#> # A tibble: 5,652 × 17
#>    dates               service  statio…¹ stati…² stati…³ MeanT…⁴ MinTe…⁵ MaxTe…⁶
#>    <dttm>              <chr>    <chr>    <chr>   <chr>     <dbl>   <dbl>   <dbl>
#>  1 2022-04-01 00:00:00 meteocat C6       Castel… Lleida      7.3     1.4    13.2
#>  2 2022-04-01 00:00:00 meteocat C7       Tàrrega Lleida      6.2     0.3    12.2
#>  3 2022-04-01 00:00:00 meteocat C8       Cervera Lleida      4.8    -0.7     9.7
#>  4 2022-04-01 00:00:00 meteocat C9       Mas de… Tarrag…     7.7     5      10.7
#>  5 2022-04-01 00:00:00 meteocat CC       Orís    Barcel…     4.6    -1       9.4
#>  6 2022-04-01 00:00:00 meteocat CD       la Seu… Lleida      2.4    -2.2     6.8
#>  7 2022-04-01 00:00:00 meteocat CE       els Ho… Barcel…     8.2     2.3    13.6
#>  8 2022-04-01 00:00:00 meteocat CG       Molló … Girona     -2.5    -5.9     1.3
#>  9 2022-04-01 00:00:00 meteocat CI       Sant P… Girona      2.6    -2.2     7.1
#> 10 2022-04-01 00:00:00 meteocat CJ       Organyà Lleida      4.8     1.1     9.2
#> # … with 5,642 more rows, 9 more variables: MeanRelativeHumidity <dbl>,
#> #   MinRelativeHumidity <dbl>, MaxRelativeHumidity <dbl>, Precipitation <dbl>,
#> #   WindDirection <dbl>, WindSpeed <dbl>, Radiation <dbl>, geom <POINT [°]>,
#> #   elevation <dbl>, and abbreviated variable names ¹​stationID, ²​station_name,
#> #   ³​station_province, ⁴​MeanTemperature, ⁵​MinTemperature, ⁶​MaxTemperature