Averages area weather

Averages the weather data series of points or grid pixels.

averagearea(object, na.rm = TRUE)

Arguments

object

An object of class SpatialPointsMeteorology-class, SpatialGridMeteorology-class or SpatialPixelsMeteorology-class.

na.rm

Boolean flag to indicate that missing values should be excluded from averages.

Value

An object of class as the input SpatialPointsMeteorology-class with weather series corresponding to the spatial average of the input.

Details

Assumes that all points/pixels represent the same area.

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

data(examplegridtopography)
data(exampleinterpolationdata)

#Interpolation of meteorology over a grid for two days
ml = interpolationgrid(exampleinterpolationdata, examplegridtopography,
                       as.Date(c("2001-02-03", "2001-06-03")))
#> Warning: `interpolationgrid()` was deprecated in meteoland 2.0.0.
#> ℹ Please use `interpolate_data()` instead.
#> ℹ Spatial_*_Topography and MetereologyInterpolationData classes are soft
#>   deprecated. Interpolator should be created with create_meteo_interpolator(),
#>   and spatial objects to interpolate should be from sf (vector) or star
#>   (raster) classes. Interpolation is performed with interpolate_data()
#> Warning: PROJ support is provided by the sf and terra packages among others
#> Interpolating day '2001-02-03' (1/2) - 
#> Warning: PROJ support is provided by the sf and terra packages among others
#> done.
#> Interpolating day '2001-06-03' (2/2) - 
#> Warning: PROJ support is provided by the sf and terra packages among others
#> done.
#> Warning: `SpatialGridMeteorology()` was deprecated in meteoland 2.0.0.
#> ℹ Spatial_*_Meteorology classes are soft deprecated. Meteorology objects are
#>   now normal sf points objects
#> ℹ The deprecated feature was likely used in the meteoland package.
#>   Please report the issue to the authors.

#Call averaging function
pa = averagearea(ml)
#> Warning: `averagearea()` was deprecated in meteoland 2.0.0.
#> ℹ Spatial_*_Meteorology classes are soft deprecated. To average meteo data
#>   series is better to use the corresponding functions in the sf package
#> Warning: `SpatialPointsMeteorology()` was deprecated in meteoland 2.0.0.
#> ℹ Spatial_*_Meteorology classes are soft deprecated. Meteorology objects are
#>   now normal sf points objects
#> ℹ The deprecated feature was likely used in the meteoland package.
#>   Please report the issue to the authors.

#Spatial information
pa
#> SpatialPoints:
#>   coords.x1 coords.x2
#> 1    364950   4635950
#> Coordinate Reference System (CRS) arguments: +proj=utm +zone=31
#> +datum=WGS84 +units=m +no_defs 

#Weather data
pa@data[[1]]
#>            DOY MeanTemperature MinTemperature MaxTemperature Precipitation
#> 2001-02-03  34        7.036145     -0.6158631       12.01121             0
#> 2001-06-03 154       21.862234     12.2691998       28.09929             0
#>            MeanRelativeHumidity MinRelativeHumidity MaxRelativeHumidity
#> 2001-02-03             79.84105            57.12066                 100
#> 2001-06-03             58.34324            40.25151                 100
#>            Radiation WindSpeed WindDirection      PET
#> 2001-02-03  10.71743  1.388889           220 1.046231
#> 2001-06-03  28.41855  3.611111           240 6.411385