Complete missing meteo variables

Calculates missing meteo variables

complete_meteo(meteo, verbose = getOption("meteoland_verbosity", TRUE))

Arguments

meteo

meteoland meteo data

verbose

Logical indicating if the function must show messages and info. Default value checks "meteoland_verbosity" option and if not set, defaults to TRUE. It can be turned off for the function with FALSE, or session wide with options(meteoland_verbosity = FALSE)

Value

the same meteo data provided with the the variables completed

Details

This function takes a meteo object (with meteoland names) and complete any missing variable if it is possible

Examples

# example data
data("meteoland_meteo_example")

# remove MinRelativeHumidity
meteoland_meteo_example$MinRelativeHumidity <- NULL
# complete vars
completed_meteo <- complete_meteo(meteoland_meteo_example)
#> ℹ Completing missing variables if possible:
#> • RelativeHumidity
#> • MinRelativeHumidity
#> • MaxRelativeHumidity
#> • Radiation
#> ✔ Done
# check MinRelativeHumidity
completed_meteo$MinRelativeHumidity
#>    [1] 44.54373 43.92829 48.22623 67.79254 48.13884 52.58872 46.28775 58.60228
#>    [9] 51.51611 56.84501 53.64692 48.23087 59.62648 49.58845 57.88316 53.23507
#>   [17] 53.84621 53.40034 48.91317 52.92733 68.74477 52.08616 68.03216 56.87095
#>   [25] 60.39051 52.51741 62.26776 40.07248 54.95452 53.07458 57.86287 55.65242
#>   [33] 64.69792 43.33776 38.93237 57.42919 50.29339 51.20680 75.15707       NA
#>   [41] 43.96618       NA       NA 50.02883 52.23688 59.78000 50.26596       NA
#>   [49] 71.57516 70.23570 52.36338 54.59605 55.16370 62.14800 62.55405 57.91424
#>   [57] 45.91763 51.83501 54.45964 48.19865 48.22326 55.10691 48.67579 44.52462
#>   [65] 54.85522 33.81187 35.77014 45.24607 50.67401 60.70006 67.35221 64.69792
#>   [73] 61.93203 42.02909 45.15919 45.74066 51.67576 42.98868 49.30001 56.32064
#>   [81] 54.95100 63.97702 48.16529 51.76176 50.73038 50.31282 47.23062 36.10267
#>   [89] 61.36378 47.40498 54.67595 60.41107 52.05833 58.66030 52.33391 53.94614
#>   [97] 68.26719 70.62465 52.76273 49.02359 48.77586 46.25286 39.17402 45.04157
#>  [105] 48.30832 57.85276 39.18674 45.87876 53.42782 56.63617 46.35729 47.64460
#>  [113] 58.77477 68.16281 50.90309 53.15720 53.53502 52.97218 47.96694 43.39702
#>  [121] 59.72323 50.12385 48.01965 48.55139 48.39733 48.97666 56.52907 47.30072
#>  [129] 55.86220 62.51645 43.19242 53.40399 52.91206 47.70686 50.00823 52.13863
#>  [137] 50.57424 52.62297 56.01142 55.21709 56.03252 54.68062 48.71304 32.62901
#>  [145] 57.54887 40.97030 48.82751 51.62347 53.04135 59.13214 48.49328 53.60259
#>  [153] 46.85083 48.90796 52.69351 54.43472 41.89553 49.75447 69.66773 55.63186
#>  [161] 57.34472 53.59226 55.95172 49.39733 48.34768 47.11483 47.43683 55.11064
#>  [169] 55.21654 49.66645 48.99867 56.90364 53.35517 53.73060 70.24330 54.88152
#>  [177] 40.36087 55.67871 60.51852 53.68215 55.82273 55.10196 69.13523 53.80061
#>  [185] 54.74703 53.13507 51.75223 51.35078 56.52169 38.70997 36.54370 39.56877
#>  [193] 62.02698 39.29574 52.89552 36.35057 54.43609 39.95525 51.23795 37.87298
#>  [201] 34.15581 46.31157 37.61715 55.21746 48.92727 40.92882 39.65953 38.20394
#>  [209] 42.33635 61.08768 37.65515 69.32789 49.33902 54.06657 45.53719 56.60999
#>  [217] 35.92147 63.09550 44.71784 40.33641 46.14801 55.05750 39.39941 35.10669
#>  [225] 59.84521 45.39361 32.66029 55.47544       NA 36.96080       NA       NA
#>  [233] 33.18379 38.55924 49.97927 38.16616       NA 56.53589 55.92352 47.23062
#>  [241] 40.84002 48.59364 55.82765 59.18653 52.98205 40.27611 39.65953 40.84002
#>  [249] 41.03482 42.86134 47.98749 44.12082 37.62257 45.49594 34.66782 35.93741
#>  [257] 39.54711 46.22348 55.70498 60.47940 53.70742 57.83715 33.74214 33.56975
#>  [265] 35.29458 35.84605 38.57781 38.52905 50.12399 50.00802 55.82133 38.94538
#>  [273] 40.24052 42.62800 41.83564 38.15018 34.74295 48.32527 37.77789 39.19004
#>  [281] 82.32867 37.99351 39.59904 37.34123 44.83761 58.37142 59.36103 38.75980
#>  [289] 39.32614 37.05237 38.83804 34.09521 34.16448 42.04500 51.77899 33.26076
#>  [297] 35.98192 46.57490 43.85776 33.11436 40.56837 38.48589 59.24216 42.57988
#>  [305] 42.55873 47.95010 43.23401 42.98868 37.02144 47.16749 50.72618 34.87881
#>  [313] 39.77030 42.29761 38.15746 43.64706 38.82436 43.61793 55.27165 35.85311
#>  [321] 48.88464 42.47827 36.54833 42.90921 39.91312 37.89243 43.94521 48.96211
#>  [329] 42.76628 41.00525 55.40882 44.80626 32.99773 43.44436 32.57189 39.41286
#>  [337] 45.30443 41.07976 51.83501 39.11095 41.02712 40.93030 38.82232 41.73673
#>  [345] 42.72757 36.24685 39.54711 57.83073 53.95898 41.23026 48.33708 60.49837
#>  [353] 36.86973 45.86951 37.35576 36.29712 50.44380 57.56335 40.44405 39.56329
#>  [361] 39.29283 46.03853 42.43181 62.93753 43.32269 35.44631 42.83709 62.50598
#>  [369] 73.27667 65.03379 86.28305 76.67442 72.15026 67.47875 54.17741 85.54885
#>  [377] 67.82470 78.18267 29.32739 34.59092 38.99102 53.83761 33.51736 31.06428
#>  [385] 48.16529 47.50133 34.28930 37.26488 32.94664 32.90709 36.04890 40.23471
#>  [393] 32.18340 40.68701 59.52181 33.40888 39.69210 70.61768 67.34253 69.57920
#>  [401] 67.81644 31.33750 42.84589 41.74154 47.85105 46.53161 45.40789 57.72231
#>  [409] 42.53140 56.74651 57.88247 44.65738 64.28463 28.82372 46.70732 30.29003
#>  [417] 64.07557       NA 56.89796       NA       NA 28.06305 27.60299 41.43034
#>  [425] 31.65971       NA 60.88071 58.21685 51.75095 44.41961 55.58222 46.06155
#>  [433] 55.14921 47.93735 63.27472 66.07943 74.77971 67.08047 46.39951 43.05460
#>  [441] 44.20864 62.17452 37.08441 58.76306 44.24054 65.74280 63.19882 57.35663
#>  [449] 55.69581 56.95586 45.09989 30.36985 28.82892 34.51514 31.51437 36.68686
#>  [457] 30.01373 42.86579 36.40371 44.23144 34.92925 28.10603 31.20332 38.41833
#>  [465] 29.29761 44.27043 41.47922 34.14759 35.03310 64.43010 49.70402 51.65843
#>  [473] 62.98227 40.48169 62.12354 59.73887 51.87304 50.44776 36.50435 35.41828
#>  [481] 28.55099 29.51931 33.21991 44.07005 55.78835 33.21991 34.09521 32.72051
#>  [489] 28.36621 35.60190 46.01510 66.53464 40.98935 51.52592 47.10971 31.26330
#>  [497] 52.61101 58.96762 76.44728 32.24183 35.08167 39.66995 58.62836 30.93297
#>  [505] 79.86411 32.88095 84.33841 44.87641 30.57149 36.36232 77.04975 51.56259
#>  [513] 41.21057 40.65531 54.31294 34.43960 45.91763 58.16766 65.07463 56.00367
#>  [521] 31.81671 37.92246 70.64117 65.53867 32.17742 26.38150 41.42388 39.05040
#>  [529] 26.69143 30.47378 50.02026 32.36141 53.09680 60.00614 30.49033 29.24432
#>  [537] 55.60887 45.00358 39.12669 41.65179 40.95254 33.26456 36.47928 31.34161
#>  [545] 35.27405 52.44617 49.63029 41.33255 32.56989 55.95001 47.17276 72.01985
#>  [553] 41.70411 49.82891 58.89107 83.13432 37.19917 51.29240 43.27081 54.48495
#>  [561] 49.67295 61.87755 51.77511 51.71069 62.44871 54.32910 46.10964 37.20307
#>  [569] 42.60924 42.36939 55.66210 36.50654 28.12229 44.98266 40.73488 39.68483
#>  [577] 33.89321 35.07644 31.29328 37.23674 35.15160 24.09920 43.02582 44.24054
#>  [585] 34.50138 36.74814 44.45524 57.73408 55.81713 55.52887 32.00615 46.61955
#>  [593] 43.82827 52.25272 44.28340 32.30634 45.94706 46.29890 57.03403 60.43086
#>  [601] 37.80844 52.52879 24.03043 46.29890 34.98563 48.91317       NA 35.36104
#>  [609]       NA       NA 32.17995 33.14861 33.07439 31.81263       NA 45.98053
#>  [617] 50.29339 42.60924 39.81050 58.16563 44.21994 57.33068 55.60887 53.13678
#>  [625] 51.51249 56.19087 44.60264 43.04048 37.77701 41.41865 45.60184 40.11851
#>  [633] 44.60264 35.97448 48.39456 46.92982 63.30485 60.57937 62.14762 50.15227
#>  [641] 35.67071 37.29372 41.93147 35.43619 39.25327 35.07644 39.12648 34.02641
#>  [649] 37.59343 40.39110 30.50386 31.35322 36.38825 33.23902 34.84857 34.99025
#>  [657] 38.11410 54.00196 52.67739 40.63239 39.88277 45.58367 48.13595 43.23401
#>  [665] 44.34235 43.61044 37.32548 38.73346 40.95918 34.16779 33.26914 37.90859
#>  [673] 45.51905 36.03509 38.18770 34.00422 35.57795 35.30598 39.53849 41.18003
#>  [681] 52.98184 41.23013 45.19079 49.73932 34.50629 43.96774 39.44244 56.45246
#>  [689] 37.20932 35.95170 43.07239 48.82751 42.66905 56.03252 39.65953 60.28191
#>  [697] 48.65672 36.02866 38.61957 53.72876 39.08068 44.74838 37.27040 41.52975
#>  [705] 37.64127 61.88382 41.81122 41.95043 36.84037 37.84806 34.35581 55.13827
#>  [713] 50.26521 40.81274 26.38582 44.18887 42.72879 28.44078 41.74385 41.80130
#>  [721] 35.57935 39.58059 43.18932 37.35576 30.11424 48.68087 39.36468 42.65778
#>  [729] 46.20039 33.32297 41.41178 32.84083 39.38273 42.33933 57.88857 25.32749
#>  [737] 45.50372 34.93924 47.78966 54.57577 47.96313 42.52905 38.27836 49.44254
#>  [745] 53.43132 34.36509 38.14455 36.28099 35.66763 42.44288 43.73652 52.51799
#>  [753] 41.18312 40.05645 44.08868 49.42094 26.65033 34.37488 33.92340 40.76199
#>  [761] 25.06563 20.93434 32.57429 35.51838 26.69143 25.29857 22.59549 27.53818
#>  [769] 34.68651 34.14613 21.16587 30.53360 40.21068 32.68906 24.95207 38.15778
#>  [777] 51.08156 52.95299 56.55501 28.27318 36.07250 41.85407 42.03559 29.60585
#>  [785] 36.88376 49.67675 34.04170 45.91711 52.76302 32.60388 43.54381 17.07366
#>  [793] 36.53015 23.14273 30.26345       NA 27.83037       NA       NA 23.73401
#>  [801] 31.83148 23.99132 22.94947       NA 35.41898 40.42879 35.26919 28.45284
#>  [809] 44.95738 34.05963 50.17623 43.52904 33.01607 32.54036 56.34588 45.10274
#>  [817] 34.43580 33.62110 36.01212 38.28318 31.90968 29.54708 26.25774 57.44796
#>  [825] 50.34971 61.77776 45.56621 53.36291 32.09569 31.56273 28.33590 26.70060
#>  [833] 25.25272 29.28228 27.65715 36.14053 26.32153 38.50344 32.06256 24.70921
#>  [841] 28.93321 31.97628 26.37271 23.72431 34.28477 32.47871 37.23488 55.00341
#>  [849] 27.21999 30.14391 39.73118 35.72081 41.39331 38.46708 36.20887 28.19412
#>  [857] 29.55249 24.89528 27.56948 29.58760 29.35166 35.54777 26.73684 34.40490
#>  [865] 28.74283 30.12881 30.83062 35.05998 41.11045 52.08616 29.27257 34.60374
#>  [873] 45.67332 28.91148 38.79407 39.13773 60.38081 21.82278 22.39934 35.18054
#>  [881] 49.99518 33.26076 53.55194 32.62855 67.73660 34.57625 26.30392 31.03806
#>  [889] 49.10303 29.22336 44.27781 29.52284 34.28710 33.05811 49.13518 32.04002
#>  [897] 31.36326 23.82834 29.62890 24.01396 49.25819 42.56639 28.94908 24.17658
#>  [905] 26.64625 34.75949 25.22428 27.49122 32.19292 25.69830 35.85923 33.34448
#>  [913] 27.86411 25.90717 34.16795 36.99112 32.19292 34.96976 24.51543 25.16740
#>  [921] 25.87859 25.82608 27.69112 48.67855 21.76315 43.87747 27.71390 38.10433
#>  [929] 44.23969 38.31795 24.95173 26.48761 43.48518 55.13718 34.87733 42.20901
#>  [937] 49.08204 67.98319 48.31063 59.80460 65.09298 49.14084 55.92418 63.55267
#>  [945] 56.04848 36.35652 43.07118 44.38585 50.03969 38.56318 29.82310 39.01687
#>  [953] 57.33511 36.93115 31.67166 33.91167 28.81274 36.28707 42.69806 25.86237
#>  [961] 38.64561 50.00802 34.39805 37.21449 48.10487 51.02649 53.07589 71.60357
#>  [969] 33.69337 42.53213 43.82152 48.24525 41.12924 67.90040 43.27521 50.01188
#>  [977] 56.07075 59.95676 38.60987 48.53872 33.44962 40.18323 34.70281 45.34034
#>  [985]       NA 39.31571       NA       NA 34.83127 33.90731 34.78956 33.73387
#>  [993]       NA 36.81964 44.48176 43.10038 41.73633 40.74393 38.09740 52.27832
#> [1001] 45.73806 44.90868 45.44992 51.64966 42.46759 43.15871 48.91938 45.10068
#> [1009] 47.42317 41.35526 40.23218 40.33981 55.94055 57.29333 54.85191 58.85046
#> [1017] 62.62220 46.66360 32.95650 36.75271 36.44785 34.48429 43.81006 37.65319
#> [1025] 40.71465 31.83197 45.48341 43.53044 33.57597 37.74339 39.88929 35.81877
#> [1033] 34.84214 42.33635 45.74322 41.17287 73.73816 36.49463 40.70874 53.56764
#> [1041] 41.22381 42.11775 43.27521 43.30429 31.99075 36.48578 39.26753 35.16520
#> [1049] 33.96635 38.27615 42.14699 39.61126 39.65314 31.00110 33.04834 34.45730
#> [1057] 41.47320 59.78890 60.11219 38.81693 43.03404 52.36264 34.81191 51.74345
#> [1065] 49.47288 61.99189 31.16718 34.42768 39.13773 48.70780 37.18977 54.77107
#> [1073] 41.73772 64.81498 45.51042 32.95650 40.62673 56.35549 45.13495 52.95470
#> [1081] 40.02447 42.40944 41.35239 52.05896 46.37784 43.70984 28.24399 37.47744
#> [1089] 37.12560 65.68219 38.06740 37.10114 31.51100 41.23707 42.94665 25.98171
#> [1097] 38.33872 46.37784 34.72303 43.73436 48.45755 35.30033 35.01807 40.42110
#> [1105] 51.31757 44.15554 40.74393 46.22953 38.92050 26.44945 34.21333 40.79784
#> [1113] 38.65944 33.07848 51.65116 37.56899 47.79450 42.08965 49.12869 34.05404
#> [1121] 37.59577 44.16158 58.95008 56.42504 66.38178 62.26871 81.76675 76.57313
#> [1129] 73.64226 64.40113 60.81066 67.98319 74.84232 71.35294 31.00870 37.00420
#> [1137] 43.68098 45.25329 29.79944 26.62041 34.86971 46.51993 34.78261 22.22225
#> [1145] 24.09621 27.66423 30.74626 33.19387 20.96818 45.40777 45.62262 28.00344
#> [1153] 25.07853 35.92475 35.59897 41.37391 62.73866 30.56877 36.34553 27.78103
#> [1161] 46.86410 28.63188 50.78378 37.86565 40.16905 32.44192 57.70239 26.22833
#> [1169] 33.28296 22.17283 32.63730 25.31299 41.09415       NA 22.11862       NA
#> [1177]       NA 22.87862 29.32160 27.17971 22.67013       NA 33.34078 43.88097
#> [1185] 27.35948 35.51077 28.45833 41.63506 45.56301 47.68497 30.67456 31.60937
#> [1193] 45.56906 34.67165 46.70260 40.03738 42.47005 34.15738 37.80713 27.97228
#> [1201] 24.92393 31.41213 37.37179 48.02586 41.84598 51.87191 34.62697 29.13450
#> [1209] 23.88804 28.66847 27.83057 33.90731 26.17991 41.05234 28.85499 41.49892
#> [1217] 37.34243 27.62877 26.57418 29.15837 25.35491 23.46979 30.58534 36.20853
#> [1225] 29.02979 44.26933 28.68917 29.86701 43.95240 31.93944 35.72997 43.63162
#> [1233] 30.51185 21.63381 32.42050 29.58023 22.97140 27.27952 35.82062 35.86657
#> [1241] 24.56496 29.30748 37.75692 26.84439 25.60727 40.65067 44.40455 63.95375
#> [1249] 25.64717 32.54351 50.20878 32.89776 33.46096 38.84646 38.79567 23.10392
#> [1257] 25.49531 34.02134 34.31484 33.00786 35.03782 32.02680 39.02682 42.26317
#> [1265] 22.33071 36.50459 35.81509 26.83987 45.13495 36.56321 27.43188 40.38817
#> [1273] 34.12837 29.24464 28.74321 25.03301 38.70985 25.45269 53.67422 31.02434
#> [1281] 31.60344 32.10537 28.72620 38.94208 23.52466 38.63949 30.23384 28.54775
#> [1289] 27.78103 29.29301 24.51188 24.17536 40.98202 40.83444 41.16528 38.47842
#> [1297] 29.89929 34.20209 24.59514 40.40448 34.50888 45.94875 28.23480 41.70839
#> [1305] 30.09093 34.84705 44.41897 32.60831 24.32782 25.18960 31.82675 42.95952
#> [1313] 40.96242 60.35042 43.53207 49.74693 42.01769 48.29593 62.98227 51.56259
#> [1321] 51.11494 59.41874 52.47086 36.22866 50.89288 57.68393 51.44512 34.59735
#> [1329] 32.98959 52.56759 42.74222 29.63419 29.31337 36.15650 37.12135 38.74759
#> [1337] 45.88178 27.03250 61.22877 54.85191 35.10627 34.63653 46.61342 37.36251
#> [1345] 51.67554 66.54968 31.95027 49.15597 37.09019 57.11018 26.52891 48.77586
#> [1353] 51.71455 31.38435 40.14187 63.88650 26.07867 32.75992 24.41732 59.41912
#> [1361] 49.70900 36.17034       NA 27.46002       NA       NA 32.76111 37.38423
#> [1369] 29.13450 28.59381       NA 27.34742 34.59021 41.78227 52.13976 49.57381
#> [1377] 54.24857 56.77193 53.96547 25.37457 28.20833 57.54793 37.78602 57.70239
#> [1385] 46.17680 56.27806 44.62716 61.03865 28.48672 24.08673 39.55249 47.16454
#> [1393] 51.92969 43.21401 62.44556 44.02343 42.93490 32.11424 32.01066 32.72799
#> [1401] 40.42110 37.08449 53.50644 34.89290 46.19270 52.11281 40.52002 35.89227
#> [1409] 34.89155 32.94047 21.17715 36.87104 44.03956 42.04675 55.95949 50.26152
#> [1417] 48.35725 47.77717 39.73833 30.83933 31.28268 42.24233 27.77340 55.60666
#> [1425] 28.40252 30.55611 35.75007 41.72450 48.74707 31.31155 41.77699 48.88547
#> [1433] 39.70941 34.40270 51.57950 52.52751 69.77172 48.63337 47.62261 58.31816
#> [1441] 50.97539 48.32223 39.54291 49.67980 25.86345 34.28554 48.98934 33.52750
#> [1449] 46.39033 40.88037 42.12069 50.30155 42.70296 31.15515 49.16100 44.41897
#> [1457] 32.65981 56.04474 52.24844 38.42061 57.68393 55.23815 41.47127 37.73325
#> [1465] 30.06318 49.56980 24.29814 60.52995 32.04789 40.11470 34.18816 35.05748
#> [1473] 52.11410 31.13920 55.18007 31.67068 41.48108 36.34461 40.70689 36.09776
#> [1481] 30.48308 33.17016 56.37610 34.92790 49.78892 34.46551 28.42961 30.10850
#> [1489] 54.87662 42.94021 51.15196 34.64355 55.04078 35.23414 46.47477 54.19753
#> [1497] 40.79344 29.07679 34.66571 33.37014 51.46112 51.89931 72.08175 49.52577
#> [1505] 56.18607 46.70546 57.24049 76.10235 59.22391 61.33882 54.57051 66.86599
#> [1513] 39.28042 48.03023 48.51383 53.50308 43.01275 41.01149 59.09892 56.00224
#> [1521] 53.48591 35.65575 40.86090 35.45808 40.30043 48.76330 28.06305 43.79248
#> [1529] 56.94885 46.98577 60.60645 59.71698 76.32362 68.36293 70.61262 48.53220
#> [1537] 57.25225 52.17035 52.40830 59.74930 52.44736 51.44401 70.12055 51.02923
#> [1545] 56.62597 58.32806 64.86791 28.91964 50.60724 48.28191 66.57087       NA
#> [1553] 58.37840       NA       NA 34.56082 39.62935 47.52827 33.40672       NA
#> [1561] 68.71392 69.64903 53.33181 46.42898 48.64300 49.05099 53.64891 50.62374
#> [1569] 69.54878 70.61334 58.30286 55.42592 47.73327 52.04929 44.27824 48.29183
#> [1577] 50.67690 54.67064 57.75511 59.54556 65.08241 50.94845 55.81211 53.88501
#> [1585] 49.87366 38.60062 38.09641 46.91127 39.77133 44.39463 35.05748 48.87791
#> [1593] 44.11236 52.05885 44.38674 34.19374 34.53838 37.86031 35.89180 51.37820
#> [1601] 42.11887 43.19754 53.64891 71.55801 60.05477 64.20480 61.23500 50.07020
#> [1609] 66.45065 68.23962 52.67817 36.02559 51.46502 53.76132 33.19334 34.37848
#> [1617] 39.56039 54.16649 55.52101 41.34070 47.42317 49.18368 31.33551 51.49227
#> [1625] 51.70650 62.75018 51.98329 56.31152 52.55761 46.25077 41.55842 62.57541
#> [1633] 67.33929 36.33707 41.51561 43.83723 52.54106 49.37779 66.52831 41.10669
#> [1641] 68.71171 55.09583 37.62790 51.15365 67.63465 57.48148 49.13518 51.62455
#> [1649] 68.36371 53.68639 54.87662 64.59061 66.81539 46.84344 46.88297 65.65828
#> [1657] 73.12039 54.74485 36.59270 35.93571 50.37164 46.92888 37.08449 47.85602
#> [1665] 58.18543 44.24235 53.43780 68.32242 40.86090 33.22609 70.13009 48.25550
#> [1673] 70.59294 53.39672 39.02267 54.40211 35.72997 40.35894 43.30429 50.62374
#> [1681] 38.66074 48.09005 41.28944 64.22719 50.71630 67.38224 40.45682 56.81647
#> [1689] 60.57866 63.52461 67.98789 66.65935 56.66014 71.51716 65.30483 68.80302
#> [1697] 69.23846 61.71909 69.28702 62.41419 59.45277 38.33853 55.80987 45.94603
#> [1705] 78.20430 50.06748 30.14618 50.63352 47.90538 53.12940 37.15524 41.21194
#> [1713] 37.10114 48.53872 48.99033 21.62712 58.19125 52.23463 45.53612 43.66087
#> [1721] 51.65183 74.20913 59.62631 69.96046 48.70780 45.24769 50.97134 76.28605
#> [1729] 56.18294 41.89614 62.79045 54.74485 67.12716 76.66302 43.72815 62.06292
#> [1737] 19.76982 64.92477 42.74222 56.45375       NA 53.77416       NA       NA
#> [1745] 37.80119 60.03324 48.48222 43.48518       NA 47.85109 54.16649 52.00949
#> [1753] 49.55691 68.44600 61.21461 68.82311 63.15098 57.73436 65.15833 60.79761
#> [1761] 47.91793 71.02173 38.80025 64.89063 57.14628 45.12982 61.84931 53.08812
#> [1769] 67.57792 63.10412 64.45508 65.28370 76.59807 69.50921 38.30888 42.03023
#> [1777] 42.44430 48.76402 49.13518 37.47744 64.20480 44.13238 49.15927 48.23175
#> [1785] 31.94457 37.21449 45.19270 42.59061 48.76402 40.91645 46.99503 71.95113
#> [1793] 44.73923 52.95299 38.01167 63.60283 69.87997 54.53497 62.13377 53.45313
#> [1801] 38.27921 68.02767 50.97996 37.00218 40.73374 41.56151 53.26462 50.41258
#> [1809] 54.00713 47.48205 51.53932 34.67318 61.23878 55.44869 67.34253 53.55194
#> [1817] 54.26876 58.77280 59.48693 42.85499 71.20373 76.25813 53.57887 42.61983
#> [1825] 48.28390 58.03419 73.13928 63.69380 48.48222 75.82582 70.26084 40.35056
#> [1833] 42.91749 64.31772 46.22901 57.04391 55.82133 47.67704 67.60177 68.82311
#> [1841] 50.35716 56.53549 26.74591 35.18150 51.78796 65.65915 44.52019 41.01149
#> [1849] 28.67215 49.61283 44.32744 32.26789 66.47552 50.85332 39.94819 53.39937
#> [1857] 51.30112 43.64706 29.90631 55.38116 41.43974 54.16649 71.64503 27.39310
#> [1865] 49.49108 33.13504 44.15554 44.13238 78.97386 25.55496 58.29354 35.87792
#> [1873] 57.56573 80.99088 61.20874 23.11824 49.32756 67.06297 63.79878 48.72733
#> [1881] 52.86689 47.31939 54.40472 50.66882 62.63952 55.79606 52.18231 57.88316
#> [1889] 52.01275 60.28409 30.70723 42.17848 40.50931 53.15635 31.95993 32.18651
#> [1897] 44.18449 50.12023 38.78739 29.36509 32.12106 33.08266 39.98087 37.39616
#> [1905] 25.39821 46.52093 48.54194 59.44354 33.63830 46.89100 72.78841 61.69421
#> [1913] 79.04825 55.52931 60.32272 40.60904 66.05351 61.83658 52.77014 41.74015
#> [1921] 54.87578 59.37321 64.74787 40.19847 68.69342 27.31256 46.50927 30.17988
#> [1929] 47.66425       NA 59.84735       NA       NA 28.40398 38.45333 33.78911
#> [1937] 29.30529       NA 45.79472 57.50685 38.30345 38.00787 56.76264 53.47652
#> [1945] 55.52101 69.54878 59.87181 58.82804 59.53672 46.45748 42.03155 41.55944
#> [1953] 42.80776 46.59422 49.07274 63.61524 52.99526 57.51475 53.68639 59.42255
#> [1961] 52.41369 73.45127 44.50195 33.54606 35.83746 33.60671 32.04769 36.23815
#> [1969] 31.32673 60.89527 56.64362 64.29721 37.98033 34.51491 38.82607 42.69618
#> [1977] 38.24479 55.70326 38.09626 35.47824 52.85123 60.16298 37.39196 37.12129
#> [1985] 45.97071 46.91934 49.12869 62.94950 40.80244 27.88951 45.96623 41.80176
#> [1993] 27.49011 36.34553 33.31371 61.69421 39.24635 39.27398 48.63041 51.70964
#> [2001] 30.01305 44.02636 44.62666 62.60149 38.39432 42.35323 59.56311 49.18342
#> [2009] 49.93922 53.74751 74.06770 37.83076 28.05814 41.11451 48.29411 63.73010
#> [2017] 59.92068 38.91916 74.73527 64.96302 32.31049 34.54405 61.35250 38.42383
#> [2025] 44.86997 35.90008 41.68170 56.12269 57.76887 40.55718 42.11770 29.01446
#> [2033] 31.22640 57.86008 58.83993 45.62262 33.92772 34.57813 37.91937 60.65478
#> [2041] 27.94648 47.75651 49.32419 34.23686 38.39175 45.22558 33.02785 30.79910
#> [2049] 55.38116 43.31441 54.01315 67.27628 34.61390 38.24298 29.03550 29.31337
#> [2057] 32.12106 63.86729 30.50021 48.28830 34.21977 49.12869 61.15783 50.49557
#> [2065] 36.71230 34.60524 58.70046 61.67042 56.81851 72.63093 56.40359 61.71314
#> [2073] 62.00442 75.76011 63.74045 60.61636 70.27069 64.79722 76.15449 63.01635
#> [2081] 63.93563 64.89036 89.38414 48.06166 47.67704 76.09327 49.68267 59.88308
#> [2089] 46.65077 56.86477 51.42480 53.75674 52.83350 70.73510 69.77172 83.27154
#> [2097] 66.82970 68.98708 85.97538 72.17771 91.13473 89.41628 75.46569 68.01040
#> [2105] 75.90878 85.53562 68.25868 58.69527 81.00457 84.78582 77.92669 90.18279
#> [2113] 83.74527 71.14572 38.24658 67.93473 67.59905 83.68563       NA 91.86374
#> [2121]       NA       NA 61.95469 72.86801 43.05110 43.25600       NA 78.02234
#> [2129] 71.50773 81.70531 65.32822 78.87073 76.43762 82.88923 88.40476 67.42943
#> [2137] 61.77776 89.78668 81.91427 79.26543 59.06053 76.03554 71.45325 65.17207
#> [2145] 82.10421 84.34909 86.70593 91.67384 90.12334 84.57620 88.45661 85.45003
#> [2153] 57.53221 58.08519 72.58844 47.20983 63.58483 62.69439 74.31449 76.90433
#> [2161] 77.40639 61.95469       NA 68.81614 74.71796 70.13009 88.85085 53.38375
#> [2169] 60.24151 78.32848 78.83201 61.16066 72.32320 58.85289 68.34307 79.04825
#> [2177] 72.94189 79.12226 50.21431 68.03133 78.61153 54.84963 62.80834 66.27045
#> [2185] 72.65925 87.21783 68.17733 69.79178 77.32784 67.46667 62.41545 55.86220
#> [2193] 77.79637 72.51284 80.07342 71.51646 73.70256 84.79545 84.97506 90.15313
#> [2201] 37.88984 47.85109 55.65128 75.90878 79.54445 90.77060 72.64073 89.03363
#> [2209] 89.04629 53.55614 69.72890 86.18123 75.75917 71.02173 80.82553 79.68915
#> [2217] 68.53330 75.94185 78.06865 83.44540 59.14838 65.51139 85.48221 51.93834
#> [2225] 81.91427 69.27120 56.58675 80.35335 73.18007 58.93131 77.39135 89.44029
#> [2233] 60.74874 75.74246 80.68851 61.01481 65.98694 71.46927 57.14987 83.68563
#> [2241] 75.60824 58.32806 71.64503 54.20478 64.54272 63.01914 83.48133 61.20874
#> [2249] 64.02130 70.71055 79.01854 84.48593 83.31269 57.14987 78.70687 69.26566
#> [2257] 89.00072 68.29293 85.85550 66.18377 80.64992 71.14907 83.42275 68.63770
#> [2265] 65.73365 78.33803 79.60820 73.44439 72.37961 65.90439 72.68208 90.54328
#> [2273] 69.85981 68.51246 77.17180 87.00034 79.62574 74.93742 66.30931 78.23523
#> [2281] 76.71799 76.55365 64.16219 74.53509 92.23179 74.72863 62.43346 84.71821
#> [2289] 91.88238 86.40589 81.59953 81.12711 88.71510 72.08361 92.41017 68.06985
#> [2297] 74.47987 89.94943 78.38898 76.01023 94.32053 61.06921 86.65660 62.99832
#> [2305] 72.37961 80.43792 77.92930       NA 73.26793       NA       NA 61.47936
#> [2313] 71.15896 64.30025 61.84104       NA 65.74664 70.97072 80.97718 79.41295
#> [2321] 79.60250 87.20831 91.24265 88.88441 71.43073 75.89330 87.99569 82.14305
#> [2329] 89.85690 79.19382 84.48965 78.94401 78.64206 75.07815 67.43816 85.58874
#> [2337] 92.90816 91.28262 91.86996 93.08696 88.82558 75.64001 73.92518 81.98695
#> [2345] 66.19952 77.59921 70.02063 84.24530 84.44051 88.93448 65.88229 83.19070
#> [2353] 81.53297 72.54787 70.75484 68.32404 72.19096 75.43899 85.83343 77.26814
#> [2361] 67.02171 79.41884 70.99024 69.32663 68.69935 78.83902 64.64230 67.02171
#> [2369] 62.53030 66.73945 78.46528 74.83333 91.16189 78.48049 80.59160 70.07484
#> [2377] 84.95297 78.02234 72.24774 72.32389 90.94811 70.95119 85.32025 87.01032
#> [2385] 82.14305 89.58782 79.57350 80.28500 63.12756 67.88442 79.72808 71.56529
#> [2393] 84.24530 85.53562 81.53297 89.56700 84.23246 67.82118 81.04554 88.98422
#> [2401] 60.02045 86.75917 72.78570 76.91306 76.50403 79.10749 70.31551 84.45188
#> [2409] 66.41863 70.99024 63.08066 80.76549 72.37961 70.93164 71.97011 83.69758
#> [2417] 87.61733 80.46599 71.47724 71.19800 64.72069 78.46528 75.38789 64.22765
#> [2425] 68.78193 68.61648 81.08487 73.35930 90.59370 69.81108 64.92477 77.82004
#> [2433] 64.92477 66.76115 90.67215 65.54849 83.49263 72.99711 79.30984 83.21520
#> [2441] 84.37202 80.89004 76.83206 81.58625 94.31614 83.33337 83.35898 79.71381
#> [2449] 81.16646 81.63476 87.07106 80.63541 82.38319 79.40999 82.91818 84.07276
#> [2457] 42.49473 44.37023 43.12129 44.96029 45.57284 42.46589 40.00691 60.38081
#> [2465] 48.02043 39.62256 35.30167 42.04150 45.99512 53.09226 35.32247 42.06061
#> [2473] 57.14628 41.28944 32.14180 42.43784 59.53946 59.17347 69.28052 38.21792
#> [2481] 55.78108 36.85787 55.16218 41.46385 53.07458 50.11949 44.91326 40.49116
#> [2489] 55.34327 34.72405 54.10891 38.09852 38.47707 33.05696 46.33334       NA
#> [2497] 37.61710       NA       NA 35.37978 35.21396 38.12569 36.17297       NA
#> [2505] 36.25753 40.02750 42.01276 50.68906 38.10522 45.88799 44.59705 53.41133
#> [2513] 39.75527 41.63775 55.56311 40.89543 41.73094 51.89683 46.54288 38.16707
#> [2521] 45.68093 42.20736 39.30964 61.34806 56.03926 44.41342 50.73226 52.06357
#> [2529] 44.01476 46.70725 45.99512 44.53072 47.88666 57.40524 41.71901 43.13211
#> [2537] 37.41338 44.31667 43.08523 40.33042 42.84416 42.66745 38.27225 39.55117
#> [2545] 38.77694 37.38423 65.54366 36.07001 43.72815 51.76077 40.53342 41.12605
#> [2553] 37.50076 42.72490 35.38928 36.82812 35.67109 42.90596 44.69736 44.32636
#> [2561] 46.79185 39.08998 47.77499 40.01007 43.45325 46.05165 42.70296 43.44001
#> [2569] 58.75054 35.62244 47.61159 56.24077 39.82356 52.14647 46.58548 58.86433
#> [2577] 34.58381 38.47653 49.10050 38.94208 42.52410 52.54188 43.46050 58.45699
#> [2585] 49.77089 42.78382 45.49285 52.06357 35.08035 50.73217 38.33289 39.97801
#> [2593] 40.00703 39.26381 39.12314 41.59471 39.39722 42.09177 34.90544 60.66632
#> [2601] 41.05676 38.06815 38.82607 45.90905 45.23718 44.45496 39.82979 39.99841
#> [2609] 35.89227 37.73325 37.37067 43.99767 40.37557 37.41338 42.55750 45.94077
#> [2617] 48.47670 47.45354 35.43997 41.89763 37.63129 43.89501 41.84598 32.87816
#> [2625] 51.67909 41.16186 45.27304 43.73910 47.50023 40.71465 36.57872 50.01538
#> [2633] 55.18811 48.41315 55.51918 56.87149 69.09502 60.71678 55.35670 68.35771
#> [2641] 53.92900 61.43842 65.79610 52.52774 44.29772 50.63453 50.50970 41.76697
#> [2649] 34.04055 35.42882 38.37746 51.67909 33.57539 37.75236 29.13942 47.60699
#> [2657] 52.62932 37.97036 27.13792 43.75688 49.35285 42.46589 29.82983 39.89135
#> [2665] 41.77511 59.80402 66.07948 33.26805 44.28394 37.65981 43.32462 34.01164
#> [2673] 56.13533 50.14000 43.60791 45.47873 40.55348 32.12873 37.27389 32.08529
#> [2681] 41.99584 30.46356 41.12300       NA 26.78965       NA       NA 31.97912
#> [2689] 37.72782 33.34491 30.48782       NA 30.88201 34.92191 41.03723 38.13824
#> [2697] 35.37793 39.19965 38.96801 40.41012 31.71496 33.12869 61.37149 37.38911
#> [2705] 43.15123 58.90031 58.53564 50.14000 45.36898 30.78822 29.02769 46.41780
#> [2713] 46.83631 45.80873 39.37201 44.12810 33.18633 43.01367 42.81529 29.98992
#> [2721] 31.26983 35.69924 41.31854 38.85509 30.60656 39.00001 49.18368 37.75236
#> [2729] 41.12325 37.75236 26.48400 49.85050 45.68093 34.67527 62.41013 32.23132
#> [2737] 35.33408 53.21627 37.35066 34.32851 32.42088 40.13346 41.26032 41.55086
#> [2745] 29.48867 37.74557 41.19082 46.42070 43.53916 30.06879 50.10446 34.85424
#> [2753] 34.21537 43.34531 51.04469 50.89450 57.82225 34.64901 40.17060 64.34611
#> [2761] 38.21570 46.82001 42.12369 57.02833 28.94429 26.97430 51.84491 45.03748
#> [2769] 40.33480 43.52308 46.79185 53.09452 39.78793 43.70984 49.21294 47.90561
#> [2777] 32.19152 61.76548 37.07122 34.44320 45.79512 39.22644 36.63415 32.35697
#> [2785] 33.98319 43.72815 27.15169 57.00069 39.22841 40.51879 33.34754 30.16127
#> [2793] 44.39802 35.93737 40.45052 34.01234 38.44763 37.61018 35.62296 42.75749
#> [2801] 37.58212 30.38172 50.04943 41.61317 42.12766 43.46014 29.89192 40.72188
#> [2809] 41.46385 43.59428 35.29163 30.66171 59.87181 41.44234 40.51427 42.05292
#> [2817] 39.25715 32.61939 33.78363 32.66627 47.86120 53.64842 52.68670 62.22293
#> [2825] 76.54896 63.07988 63.96714 71.08085 65.18936 71.46901 72.34107 71.56590
#> [2833] 33.23917 41.21231 42.18126 43.72582 27.42329 28.71392 36.66585 46.16443
#> [2841] 25.95907 29.39408 26.17250 30.68592 38.96911 39.63884 23.39136 40.45006
#> [2849] 45.40340 33.72391 27.94767 42.76606 55.05107 50.60079 57.96698 29.22157
#> [2857] 46.19389 33.98491 41.20865 30.70226 54.87091 52.75395 35.07666 44.07168
#> [2865] 45.70316 34.50035 40.48570 20.87984 36.86371 27.42329 40.92467       NA
#> [2873] 28.93967       NA       NA 25.28064 31.47015 28.09541 21.64446       NA
#> [2881] 41.73183 29.39043 31.28573 33.36404 34.12996 37.36482 39.28374 40.79919
#> [2889] 30.75046 35.19505 56.73609 36.90705 44.37278 44.57095 49.76159 44.05810
#> [2897] 39.82502 28.19000 25.44555 45.57140 54.55050 48.45808 43.58414 56.77741
#> [2905] 38.53617 38.04729 33.55040 27.14369 25.39059 32.73846 31.36928 38.05676
#> [2913] 30.79667 33.36404 43.24257 34.04113 34.47845 30.88513 23.84341 41.78849
#> [2921] 40.08565 32.98357 64.89002 29.41855 32.61171 49.89806 36.74911 62.97573
#> [2929] 42.43485 37.45670 32.61551 36.92713 25.37457 29.09277 31.94478 33.80928
#> [2937] 48.98051 31.47015 38.07626 34.50471 33.92745 32.12325 40.16150 57.47158
#> [2945] 57.76970 29.73284 38.74125 56.25222 33.89871 44.01476 45.77541 56.31062
#> [2953] 23.69433 24.95855 42.10398 38.36529 39.19778 45.36905 36.01194 53.93042
#> [2961] 48.79763 32.60831 39.60517 50.19248 28.41393 54.40813 34.65598 33.02919
#> [2969] 40.59130 35.69354 35.05171 32.39715 31.09783 33.12359 24.24801 57.12079
#> [2977] 37.44664 33.55304 25.75995 28.10200 42.66835 24.88532 35.01807 32.27144
#> [2985] 32.00796 36.49059 36.93574 31.64466 26.91869 31.77016 45.02685 34.21438
#> [2993] 42.16690 33.60533 25.17724 23.81255 31.71496 36.28712 43.01196 32.51442
#> [3001] 52.08726 33.46991 42.54108 41.73730 42.17356 37.43612 30.37463 33.07101
#> [3009] 51.87099 55.41094 66.77112 59.04400 71.24958 62.55608 80.43377 64.95946
#> [3017] 55.66210 63.77772 64.52047 65.45874 30.37767 37.50771 37.49455 44.98095
#> [3025] 31.05312 27.84104 40.38140 52.58961 33.64115 35.40004 30.49217 32.14544
#> [3033] 37.41338 37.84416 20.27951 40.83764 51.02889 34.89881 35.35067 43.76030
#> [3041] 52.46135 52.67502 69.16029 35.11358 38.94208 41.03967 45.80873 42.32148
#> [3049] 42.13450 46.44664 52.46178 57.34701 52.27685 37.55896 42.00835 18.80003
#> [3057] 44.81098 31.76003 49.07852       NA 36.11217       NA       NA 24.83056
#> [3065] 31.16818 30.41818 22.82422       NA 38.94477 44.93176 42.81098 40.92268
#> [3073] 42.01811 40.51427 43.63616 44.93649 48.66382 48.44892 60.08116 44.41342
#> [3081] 42.61009 42.64171 49.62644 40.83662 43.94735 37.07122 34.28554 54.61708
#> [3089] 50.04280 45.66831 43.72582 47.74023 35.21824 34.21977 29.99707 29.93975
#> [3097] 30.54060 36.54961 30.09093 42.95290 37.88385 50.16697 37.05126 29.97724
#> [3105] 33.40555 27.43188 31.84201 36.37471 40.33042 41.28758 56.59595 38.53501
#> [3113] 35.34098 47.70517 40.16150 41.25878 44.71385 42.15635 28.91577 38.65167
#> [3121] 33.51676 30.13736 30.34042 31.01173 44.58335 35.68062 36.84018 39.91159
#> [3129] 35.63173 27.31926 34.48176 45.99432 62.53201 39.83331 44.03280 62.10377
#> [3137] 40.57770 49.29956 41.60880 68.87368 25.87502 27.55018 39.74633 43.48190
#> [3145] 43.30415 62.60733 33.72239 69.01446 40.71465 28.63032 41.41308 55.99090
#> [3153] 36.43159 46.48093 35.48733 42.98149 40.28585 47.23369 39.74633 40.95134
#> [3161] 30.19748 30.47303 31.28268 68.66764 41.39311 32.22840 27.06793 35.38928
#> [3169] 43.09527 24.33932 30.99622 42.46659 30.81057 41.17228 42.49691 28.34474
#> [3177] 27.93333 46.10779 47.81749 50.29652 44.35665 29.10083 33.78009 27.24825
#> [3185] 32.67330 36.95437 39.73125 23.14138 47.63503 32.12325 46.08155 41.03967
#> [3193] 47.75043 25.43853 37.87816 47.43838 59.71170 48.81195 58.06042 53.83761
#> [3201] 69.97358 56.97723 70.30157 63.41381 52.73522 63.10387 71.12083 67.11652
#> [3209] 38.85509 42.49532 44.35500 53.55614 31.35567 36.44954 46.33395 53.28161
#> [3217] 34.62334 34.39826 29.58916 44.42650 43.92894 39.24485 31.22640 38.41843
#> [3225] 50.93544 40.70689 35.90635 47.12875 45.43615 58.73408 71.46083 37.49495
#> [3233] 49.40474 41.80596 44.71045 33.72125 47.06706 48.33765 41.95522 56.01353
#> [3241] 60.63639 39.01131 46.89100 22.49578 40.94101 30.28153 43.89030       NA
#> [3249] 36.30837       NA       NA 29.32596 31.77246 33.48781 30.45489       NA
#> [3257] 34.67165 37.48442 42.37929 37.12957 44.59956 42.08144 54.00424 45.96896
#> [3265] 34.88353 40.05203 59.19777 44.45635 37.75692 42.91749 43.82516 45.54949
#> [3273] 46.36215 37.92013 34.20209 59.47181 56.89053 60.51613 58.61858 59.63879
#> [3281] 51.42480 43.13706 40.10363 29.60192 32.60168 38.18439 36.81806 46.69712
#> [3289] 35.27618 49.86939 40.91204 38.53501 43.65350 38.53467 32.47308 41.18139
#> [3297] 34.92091 42.49473 56.84501 38.24298 38.99390 51.38315 39.75527 35.05902
#> [3305] 38.92352 38.38922 33.17016 38.41843 33.02131 35.49865 39.44239 41.64451
#> [3313] 48.54097 36.32007 45.69081 47.03871 36.56264 36.08306 45.60607 50.53719
#> [3321] 62.52860 40.82504 44.85636 70.47687 34.74772 47.97638 46.28623 67.17061
#> [3329] 31.51100 26.61126 48.46897 46.33580 40.70689 67.65556 42.35152 76.93624
#> [3337] 52.41902 38.38922 46.27798 58.03530 37.89239 46.68487 36.39679 43.07464
#> [3345] 45.12314 50.35123 45.28299 43.51800 43.30415 40.77979 33.11183 64.42311
#> [3353] 44.54232 40.21932 35.72822 32.19378 46.72722 38.39006 36.55383 41.49287
#> [3361] 33.78009 41.12325 44.65674 36.03851 36.76549 35.09687 46.57131 45.57284
#> [3369] 47.99806 31.58844 33.54619 35.22582 40.54887 38.95280 47.10836 35.37027
#> [3377] 48.41314 44.17493 52.38323 47.45478 51.65541 46.51443 33.28950 36.61298
#> [3385] 62.04884 47.99094 59.52445 50.87099 61.51477 58.35002 70.43468 62.27795
#> [3393] 62.19874 63.99072 68.87305 69.87006 53.82036 58.08468 50.84248 58.08468
#> [3401] 61.77776 60.52989 51.49227 76.46410 65.30177 47.43759 41.88604 63.33762
#> [3409] 61.38336 61.58610 56.65933 48.87625 62.76689 54.03374 48.62617 55.06169
#> [3417] 56.77050 62.04884 59.09976 54.30321 58.72472 51.11028 55.51122 43.39589
#> [3425] 61.96006 53.46398 59.41905 72.05747 68.91790 44.39895 50.47050 46.06751
#> [3433] 43.64187 35.98014 58.65981       NA 38.77694       NA       NA 47.59591
#> [3441] 42.93490 61.23878 49.19753       NA 47.25822 47.85877 43.95273 53.61398
#> [3449] 60.06682 57.11954 61.75607 60.37119 43.74741 44.24293 67.10578 43.69941
#> [3457] 44.28397 56.34588 52.22843 47.60699 54.58799 41.17287 41.39906 57.55754
#> [3465] 60.77296 67.65556 62.37594 63.50183 56.78826 59.68561 55.66389 49.04463
#> [3473] 60.94336 63.24445 50.86083 55.19828 56.01353 63.44197 57.32883 58.60400
#> [3481] 61.16066 52.75237 35.29307 63.60803 55.26958 56.14201 70.05604 49.37989
#> [3489] 48.87605 56.60729 44.05505 49.90553 49.73649 43.23740 50.97134 44.02433
#> [3497] 64.15909 49.64871 59.15022 58.72551 57.48923 42.99826 68.30044 58.35324
#> [3505] 52.41013 57.54025 60.31485 63.18735 67.43816 45.99432 50.57840 60.18358
#> [3513] 45.88186 61.22333 50.30584 70.67172 58.60007 42.23411 64.87956 58.41100
#> [3521] 53.10272 61.46692 59.93220 68.67104 62.41579 51.86944 59.37321 59.83885
#> [3529] 48.37842 56.52907 45.21873 49.60199 54.42872 70.67172 49.57880 52.35631
#> [3537] 52.16969 55.33783 46.64635 65.91529 43.84298 54.19298 59.82286 52.03349
#> [3545] 57.56573 56.32424 46.42272 43.19469 56.37302 43.20097 53.59733 47.81340
#> [3553] 52.71667 48.09711 58.29354 57.06878 52.49469 54.99515 60.27417 61.11251
#> [3561] 52.68970 54.10984 59.96947 57.31006 59.30427 58.05945 59.72474 52.35631
#> [3569] 55.78108 61.34818 40.52129 40.86090 60.04251 66.05448 71.01989 58.06300
#> [3577] 73.31111 70.95833 68.55670 63.42692 69.03109 65.65701 62.90882 67.45647
#> [3585] 61.23275 57.69284 62.40978 80.12527 62.55405 53.02088 65.99259 79.72883
#> [3593] 79.25417 55.28532 56.45113 50.32942 55.07452 66.78250 46.93740 70.21556
#> [3601] 84.33841 62.33862 56.45375 77.80437 76.54919 85.88129 63.04822 78.71812
#> [3609] 83.05878 62.66881 73.58228 61.47936 75.12424 74.79849 70.49427 62.39209
#> [3617] 71.56529 72.36082 68.27957 54.63892 55.52101 61.19043 69.98054       NA
#> [3625] 71.50773       NA       NA 52.64394 69.16120 58.19125 50.27583       NA
#> [3633] 58.99977 64.61313 66.52747 63.72376 63.41896 85.79687 78.34302 70.37526
#> [3641] 64.47772 60.74874 84.50119 72.47588 68.00656 67.53304 73.21450 60.75121
#> [3649] 66.84797 67.32280 73.85351 81.82327 74.08368 75.00647 75.87653 73.43209
#> [3657] 78.98095 57.02146 59.65095 69.32564 68.23981 73.41766 57.86998 73.04685
#> [3665] 81.70902 72.32320 59.55229 71.64503 68.49160 62.78632 71.45001 57.49935
#> [3673] 53.22732 62.48170 72.86654 60.60302 55.25573 69.30819 50.66106 62.87968
#> [3681] 59.42255 63.70717 52.71667 61.58610 79.64041 48.76402 66.50574 62.34056
#> [3689] 75.62691 62.71610 67.77893 64.34611 87.42920 65.26142 60.24151 63.36915
#> [3697] 83.20937 62.03926 68.63722 72.84257 74.95621 88.31104 74.54398 80.28500
#> [3705] 57.48607 50.19047 55.70498 62.02533 70.55666 86.28251 68.85492 86.26234
#> [3713] 79.65356 53.87382 66.41863 84.03879 57.95831 66.40819 60.10805 67.66236
#> [3721] 65.34610 59.34546 67.51439 72.25267 56.42512 61.57504 69.67112 83.63697
#> [3729] 72.45737 62.69265 67.82118 70.21556 74.15979 57.25225 73.97836 70.55353
#> [3737] 52.23463 60.29846 62.43473 57.84473 64.04520 58.62897 53.94614 64.34161
#> [3745] 84.99961 59.23408 78.59624 59.78448 61.84931 57.07288 71.08755 60.00827
#> [3753] 60.97188 60.82134 74.95471 70.61262 70.29926 59.75248 61.87312 69.30633
#> [3761] 81.36883 82.46691 78.25486 75.31621 68.80744 80.67483 79.31708 83.81243
#> [3769] 86.40163 77.23646 84.37678 80.07203 56.42789 58.80022 66.36759 52.11527
#> [3777] 68.30301 47.91943 74.60571 65.35585 67.98527 54.83590 66.49947 52.87169
#> [3785] 59.58643 77.71668 41.76553 60.70006 79.86411 59.43706 73.95776 75.65701
#> [3793] 81.61280 72.13402 66.84776 51.78796 55.69581 65.99259 52.58165 77.74153
#> [3801] 79.37925 72.53205 72.32320 75.94351 54.48318 72.36082 74.83333 52.41844
#> [3809] 65.08241 65.21186 70.10061       NA 80.85323       NA       NA 68.44982
#> [3817] 62.62518 56.29336 56.53132       NA 76.38521 73.39573 69.36656 78.00168
#> [3825] 62.20445 48.71410 52.85123 53.04903 76.26946 72.49194 74.74522 84.84184
#> [3833] 70.92316 68.40609 73.25160 81.38532 59.36211 74.57043 78.20430 76.28605
#> [3841] 79.53744 49.23905 56.18294 55.00936 48.65421 55.37515 58.37840 70.03447
#> [3849] 71.43189 78.81370 58.65031 53.45313 54.21946 57.04418 61.91293 49.66222
#> [3857] 51.89583 51.30112 77.15568 67.62325 72.28581 65.65828 80.64534 76.00186
#> [3865] 74.44643 82.91903 67.34416 66.61420 72.02694 74.11950 58.82534 64.02130
#> [3873] 69.07852 64.81159 47.91793 57.83145 53.49802 79.81521 54.74657 55.12734
#> [3881] 53.56044 46.42898 60.02045 75.89683 70.05304 72.41712 73.60049 64.70409
#> [3889] 56.29827 79.61294 79.81521 78.43481 49.71464 65.14423 60.23563 74.20874
#> [3897] 56.89257 82.15597 47.52272 77.99139 58.06040 60.09431 54.34924 67.92649
#> [3905] 75.28525 73.04666 70.45687 82.47412 63.16379 63.41896 72.26663 79.39827
#> [3913] 47.62433 57.90762 75.69096 74.48812 85.95455 54.66736 53.88056 75.98499
#> [3921] 53.89575 52.77944 69.01672 73.21290 71.52905 76.18629 82.48691 57.51475
#> [3929] 48.16107 74.79849 57.95994 73.32281 55.73123 68.57237 62.89229 53.65953
#> [3937] 58.15124 66.84797 57.43067 45.53958 71.71147 52.66272 73.65499 60.70024
#> [3945] 75.69096 62.05312 74.19093 82.43566 82.14305 75.50002 65.58853 73.08044
#> [3953] 78.25486 68.22755 75.35519 78.41740 76.29761 79.37974 82.41269 84.81893
#> [3961] 39.61126 44.87977 46.71070 44.85765 43.79248 37.03198 47.49465 50.12399
#> [3969] 38.83135 42.11775 34.38427 43.15034 52.39185 44.64792 44.67059 48.84822
#> [3977] 53.18355 38.72278 44.01038 48.60823 42.20736 52.49469 59.68741 32.50810
#> [3985] 42.91749 40.90978 50.70173 38.04090 49.22355 47.33823 39.72615 54.85191
#> [3993] 56.65998 38.27265 50.99010 25.40221 53.42334 35.55320 38.24479       NA
#> [4001] 45.79675       NA       NA 39.05279 34.59735 33.98216 37.98483       NA
#> [4009] 32.57260 35.05902 39.42340 43.00493 48.93366 39.39722 46.89100 47.29740
#> [4017] 35.30167 37.89818 61.82548 43.56535 47.00429 43.35222 47.58543 49.21124
#> [4025] 43.37275 38.47994 34.86218 54.03961 56.71135 51.95028 55.52101 56.43275
#> [4033] 40.65605 41.06438 39.12805 41.07737 42.32399 45.16077 36.11960 39.57058
#> [4041] 33.13838 41.77268 46.37784 36.87104 37.92042 35.70034 31.69667 49.41678
#> [4049] 49.56447 41.05608 75.20797 44.06797 44.71340 53.68043 42.43852 32.31798
#> [4057] 34.13226 43.32433 45.76683 47.82280 37.74203 39.60721 37.24410 42.88832
#> [4065] 41.18802 38.59735 43.43103 34.18372 37.48442 37.30328 43.50124 46.45159
#> [4073] 60.79761 50.56466 48.40649 67.75134 37.57258 47.55686 51.62455 59.88746
#> [4081] 37.24021 33.07617 45.04465 45.93916 40.49202 60.17755 43.43103 59.73596
#> [4089] 49.31881 46.05696 48.37842 56.79102 37.37179 51.23434 44.27824 38.82838
#> [4097] 42.00133 50.93544 43.14313 41.89279 42.83709 40.95997 34.75320 53.42631
#> [4105] 51.97778 41.76704 33.42899 44.02433 39.55226 48.98819 37.59343 41.78227
#> [4113] 39.74180 39.59997 46.42070 34.77192 34.96895 33.86687 51.17865 35.41846
#> [4121] 42.05940 32.68847 33.45841 45.96782 36.22766 44.56071 55.50349 38.44364
#> [4129] 50.68605 39.02335 46.74155 51.26232 46.70725 33.94646 38.27269 43.54740
#> [4137] 62.07232 54.18252 75.51804 53.90894 65.12356 70.20249 75.02609 64.59228
#> [4145] 58.93133 66.88692 67.06806 68.41236 59.98129 62.43346 64.23131 57.34957
#> [4153] 58.21685 69.99181 57.12423 58.91498 60.68316 56.84494 51.26823 61.35721
#> [4161] 62.38740 62.71236 74.26127 64.09026 67.36424 69.61054 62.03235 58.85046
#> [4169] 57.95831 57.12994 68.34212 57.75954 54.86249 53.29007 55.25062 60.17685
#> [4177] 63.54793 64.59694 63.68421 69.60898 65.17002 52.64394 71.75600 49.23798
#> [4185] 61.92069 55.91445 56.89046       NA 63.21103       NA       NA 48.22708
#> [4193] 51.02649 61.22508 55.41533       NA 46.25763 52.91447 51.70468 63.15098
#> [4201] 59.27170 56.49073 58.75880 57.70915 60.55804 61.76548 65.03112 59.22243
#> [4209] 49.71464 67.73553 63.16379 56.78826 65.39136 60.84627 64.75534 63.11811
#> [4217] 71.37338 54.20478 62.67125 63.59283 50.97454 57.79473 61.16066 53.57232
#> [4225] 53.00938 59.90816 60.65167 56.27806 56.58675 64.85962 61.93685 60.77296
#> [4233] 63.09483 62.43473 60.21703 62.78413 62.99832 61.42412 80.57133 57.25869
#> [4241] 62.86866 68.65767 63.23421 47.02550 53.84710 55.01447 60.28191 63.60803
#> [4249] 65.70481 62.19577 55.49490 63.14137 57.88225 59.47808 62.64570 65.74664
#> [4257] 59.74930 58.80054 61.63413 59.30067 67.12434 62.41122 63.86199 72.39742
#> [4265] 60.05477 59.15766 72.69624 69.32663 54.50344 55.65242 65.32352 62.32187
#> [4273] 53.53364 56.40702 58.92566 60.15248 57.17456 64.92477 63.70717 58.21057
#> [4281] 63.90791 59.58828 64.18525 54.50799 66.78250 60.84733 61.38336 66.18993
#> [4289] 66.89586 64.43242 64.43242 70.76267 58.67885 59.27429 56.05389 57.02658
#> [4297] 59.88304 63.44520 65.85551 61.43140 53.59485 59.47181 62.73953 59.00069
#> [4305] 57.22266 51.43307 67.53304 54.77677 61.16066 60.77766 58.90061 67.24433
#> [4313] 61.92069 65.50365 59.99384 67.16814 65.11936 61.71799 64.55151 54.29877
#> [4321] 67.36551 54.16483 58.47886 65.28370 70.57433 74.86079 75.60353 72.72919
#> [4329] 79.87391 74.42869 79.75058 73.06533 75.48198 75.11727 75.87383 74.52554
#> [4337] 44.05505 45.56906 48.57003 53.84669 38.33140 36.83459 42.21003 41.71164
#> [4345] 43.60237 38.51615 42.52355 37.36671 40.21255 34.45730 39.71085 47.83353
#> [4353] 42.35323 45.28797 42.35152 44.29982 45.45909 42.72621 46.10779 41.14342
#> [4361] 54.32517 42.23879 51.82818 44.26905 45.45333 48.22610 44.24037 54.53886
#> [4369] 51.24557 45.94077 50.19571 43.22058 48.06922 40.87154 42.63968       NA
#> [4377] 47.49465       NA       NA 38.62418 43.10823 31.39277 33.98216       NA
#> [4385] 39.63872 47.14553 46.87629 41.77268 41.98946 44.90829 47.83353 45.55569
#> [4393] 45.99432 42.43726 41.48650 41.73829 46.94768 40.75268 50.21830 41.09727
#> [4401] 47.45144 40.94847 43.31647 54.71893 48.29411 48.80248 54.61708 58.83984
#> [4409] 43.47566 37.36671 39.70673 39.76517 37.36671 39.88929 45.68303 50.37856
#> [4417] 39.39038 55.42060 46.33894 40.09760 45.31653 40.53414 41.91794 42.92499
#> [4425] 45.16383 41.28758 73.77639 38.62256 39.35494 39.60721 45.35097 38.98206
#> [4433] 47.77717 47.01671 43.30429 44.13988 40.60904 38.96444 37.86565 46.36412
#> [4441] 53.80973 40.30137 43.55545 43.70984 44.25538 39.73596 49.18342 36.32846
#> [4449] 45.97493 41.38445 49.23905 50.30034 45.77129 39.03918 61.51934 56.34660
#> [4457] 29.88792 39.30964 38.81115 42.32496 44.16962 47.88885 42.23411 48.23283
#> [4465] 49.74277 37.57258 49.32197 46.33395 40.25976 49.88318 46.87194 40.85107
#> [4473] 48.82017 49.54387 46.30574 43.82456 42.67822 47.05211 40.35946 39.20167
#> [4481] 37.16757 44.90322 34.81785 42.49517 52.03184 40.83167 48.08472 43.57545
#> [4489] 36.93747 39.06097 44.81098 45.37360 38.91569 42.43726 42.60945 44.31274
#> [4497] 50.01538 42.64777 38.98752 45.30800 49.23905 46.08231 48.44095 42.47827
#> [4505] 47.39647 46.27751 42.58116 47.95646 46.70260 46.51993 42.81260 46.72055
#> [4513] 57.22111 72.26238 61.58900 50.58791 68.09458 76.84903 68.69838 67.53682
#> [4521] 51.80133 67.37853 58.59116 64.85430 31.26882 40.15589 37.53952 51.15365
#> [4529] 37.62790 36.87104 41.47127 58.50014 50.88102 31.28863 30.97882 31.40058
#> [4537] 39.95165 40.22789 29.91714 42.11770 46.69208 37.20189 34.22690 40.72188
#> [4545] 56.03926 54.35571 66.39971 34.33631 43.88272 40.73792 53.68849 42.43726
#> [4553] 48.40598 47.38204 45.88186 55.27677 58.87762 39.19269 49.29460 36.56706
#> [4561] 39.88190 28.77749 56.21887       NA 36.60270       NA       NA 29.33628
#> [4569] 32.18097 44.68450 32.42617       NA 45.96591 44.45635 42.14770 39.55117
#> [4577] 50.59211 45.03748 48.25401 56.33718 44.84609 46.42070 57.01795 44.25538
#> [4585] 39.45553 47.94192 43.76343 44.46937 39.19269 40.37477 38.68884 55.83287
#> [4593] 60.96601 51.12644 54.32947 59.14892 38.77010 31.18430 31.72575 34.50201
#> [4601] 36.93747 43.36241 28.68917 38.97384 34.39452 39.21036 36.85750 30.97882
#> [4609] 31.32994 26.92456 35.45191 42.91749 35.50785 44.70298 54.18252 40.75501
#> [4617] 35.43513 52.37154 43.43172 51.81790 56.81385 44.57095 33.75634 38.33289
#> [4625] 41.62706 29.99371 31.57448 31.40058 42.35039 35.71642 38.59496 29.30748
#> [4633] 36.99303 30.63275 37.80434 51.78716 56.50543 35.84760 44.34149 55.97659
#> [4641] 38.36010 40.09657 51.62826 66.14051 38.03735 30.26275 37.48442 46.42418
#> [4649] 43.54673 59.96024 34.49587 60.99434 45.80125 31.09682 38.94477 50.74352
#> [4657] 41.31854 52.41902 35.16882 37.66958 38.57525 52.95854 38.41937 37.58803
#> [4665] 36.20853 30.64948 37.00420 59.82954 38.68259 31.42959 25.74489 34.77399
#> [4673] 44.90322 30.21541 31.15643 46.82001 32.78672 42.75749 39.99841 28.54591
#> [4681] 28.42954 44.77090 50.96430 48.93650 44.74239 43.69009 44.92294 31.59864
#> [4689] 29.48023 33.28178 49.48847 37.24021 43.96647 30.49373 44.96029 48.25401
#> [4697] 47.45478 39.29613 35.42882 44.02433 58.90238 51.78555 68.22755 54.62561
#> [4705] 62.98227 57.24318 60.80671 71.60956 58.40941 64.18331 66.16964 78.48194
#> [4713] 27.89761 36.57872 38.09740 46.45256 31.92912 30.73262 40.01672 47.02381
#> [4721] 39.06032 30.42553 26.07109 32.78209 39.71731 37.12135 25.58239 30.62059
#> [4729] 41.40577 44.51184 30.32053 38.01831 43.07464 53.70150 75.44890 37.10944
#> [4737] 41.01688 35.93571 58.01032 35.17581 44.59777 47.05211 44.79941 53.26337
#> [4745] 62.16609 31.96837 46.68487 23.72553 37.51840 26.08806 45.93852       NA
#> [4753] 38.47795       NA       NA 25.38386 32.49228 31.84281 25.31398       NA
#> [4761] 38.15508 42.47005 38.47795 33.51784 45.45483 46.58713 54.46048 55.19220
#> [4769] 40.97764 41.40577 56.16765 35.92154 45.32970 42.71295 45.85196 43.33278
#> [4777] 54.69514 34.85424 36.33246 49.85050 50.10446 51.81513 54.61434 53.30527
#> [4785] 39.95165 28.40107 27.95758 28.89222 29.18262 33.14100 26.73252 50.56938
#> [4793] 32.54351 39.74633 33.46123 30.14025 42.58142 26.15003 30.99352 37.80713
#> [4801] 33.07986 39.61896 48.05508 33.34290 30.00236 48.26597 40.75501 37.34837
#> [4809] 44.94400 39.24485 29.53776 33.90072 30.10850 26.67834 30.25468 30.01849
#> [4817] 50.89288 31.92912 35.52515 33.02427 34.12998 32.53675 38.30181 51.59725
#> [4825] 63.86729 34.31484 39.88190 56.84494 40.21932 37.10712 49.69000 59.19800
#> [4833] 29.77043 25.35612 39.42638 40.72585 49.51485 46.58713 38.97384 53.58520
#> [4841] 53.23762 27.86334 36.44648 47.15672 32.91133 49.07197 30.62980 31.68620
#> [4849] 59.88746 50.72909 34.01544 33.40861 28.08674 27.31926 35.25488 52.95470
#> [4857] 33.37655 28.51574 26.53660 31.31353 48.25401 27.44929 36.57486 36.18447
#> [4865] 28.71780 34.13521 35.50595 28.17571 28.85590 42.43852 44.08214 44.95159
#> [4873] 48.66382 31.83449 31.18190 30.69546 30.55908 29.15837 50.30584 30.01230
#> [4881] 49.40549 31.25439 41.08566 42.11868 40.66393 29.42267 31.76566 41.39331
#> [4889] 53.68849 46.72478 64.97049 49.64035 53.79092 57.79896 64.23351 65.79706
#> [4897] 56.87149 60.37430 64.89308 62.90840 37.75236 48.80868 46.88297 57.11954
#> [4905] 31.70767 38.62667 47.29631 56.50311 35.22808 34.04466 32.45620 47.10836
#> [4913] 50.21431 43.02163 37.45501 42.43726 51.23518 52.59795 35.25304 46.61342
#> [4921] 45.21632 66.69874 72.36459 53.48015 54.17164 45.43061 56.19327 37.32590
#> [4929] 54.92138 52.81704 43.36138 59.66810 66.92165 34.30719 44.55947 28.31939
#> [4937] 48.14226 31.45858 43.95756       NA 35.89227       NA       NA 29.88981
#> [4945] 39.42638 35.67109 31.30624       NA 41.80596 43.75045 46.96761 40.43154
#> [4953] 47.95751 48.38583 62.42237 55.91041 40.18501 40.34416 63.01635 41.70214
#> [4961] 46.64635 48.76402 52.25119 49.16100 60.73226 37.03350 35.46931 61.25245
#> [4969] 58.61020 62.97026 63.61553 67.67646 50.97454 46.16558 44.14100 31.56255
#> [4977] 32.59562 38.71455 38.56573 58.33353 42.75361 55.91439 44.85636 40.53342
#> [4985] 50.35123 43.57545 32.50810 43.67065 42.32126 50.83836 67.46716 37.75236
#> [4993] 37.94954 56.35549 50.07695 40.15589 44.94400 43.25225 34.04735 45.20220
#> [5001] 32.85301 33.72270 44.79941 42.87302 59.93600 34.57317 50.63453 49.84626
#> [5009] 47.94239 41.01688 53.35834 51.22628 69.91178 43.34531 49.21137 73.17733
#> [5017] 53.05707 43.93751 46.77743 74.21468 32.12793 30.60904 53.29007 48.46897
#> [5025] 54.79718 57.17079 51.07196 67.71454 60.57866 38.09641 44.22681 59.20053
#> [5033] 40.76697 52.41902 39.02335 40.67801 65.37955 52.84378 40.98784 39.79614
#> [5041] 39.80535 41.43482 32.32603 61.19899 42.38234 40.34416 34.42361 32.94047
#> [5049] 54.90126 39.92242 45.16077 38.36010 33.92772 41.32651 43.83319 35.64516
#> [5057] 38.21570 43.92388 53.77416 45.68303 61.27611 38.18553 33.48870 36.28712
#> [5065] 45.23718 38.18453 57.96032 39.61896 60.24987 42.78382 46.16452 57.19553
#> [5073] 49.05099 44.74239 37.32590 44.63558 60.42919 62.77360 66.53464 58.24105
#> [5081] 73.34797 67.29280 77.30428 75.46207 63.95872 72.28274 74.37058 76.37859
#> [5089] 43.36138 46.95490 46.19272 60.23646 35.96653 42.26361 48.43041 62.29109
#> [5097] 44.12656 38.62178 34.54405 39.88275 46.47477 41.02784 37.51382 49.51610
#> [5105] 55.83287 55.08429 37.95333 45.69643 56.39756 61.48820 61.94407 46.75518
#> [5113] 55.48524 46.47358 60.47718 42.84137 56.00224 53.13675 46.50288 66.05188
#> [5121] 66.63497 36.69504 45.82785 33.57726 51.21855 33.09007 43.88145       NA
#> [5129] 34.21537       NA       NA 32.17068 47.94239 36.18586 33.48746       NA
#> [5137] 38.24691 45.43615 48.67065 38.07277 52.95057 50.69561 59.71170 57.60196
#> [5145] 43.98616 45.65799 63.94333 46.24901 48.34731 49.80800 48.24525 47.88885
#> [5153] 49.57429 39.45553 39.42638 56.37209 52.03353 62.46872 65.77304 67.56930
#> [5161] 52.38152 44.16962 43.65350 37.14321 36.67194 47.17373 44.35500 53.26919
#> [5169] 42.95414 54.18722 43.30415 49.48735 48.79763 43.51054 36.06772 43.25225
#> [5177] 36.52553 54.32947 63.39077 40.13257 36.02371 49.79540 52.97722 39.85389
#> [5185] 42.26753 46.50288 33.21029 47.28635 38.41937 36.25789 42.95414 48.26406
#> [5193] 52.00664 38.47653 51.87601 43.96895 47.78983 45.19678 52.30110 50.75643
#> [5201] 60.41958 44.21446 49.18537 59.93853 52.73156 48.96042 46.50288 72.76963
#> [5209] 34.82495 29.92629 51.34364 54.12610 56.43962 64.14580 52.75823 74.43972
#> [5217] 60.21098 39.26845 54.25714 59.90529 39.00001 52.35960 42.14682 43.53271
#> [5225] 52.00482 63.52461 42.46589 42.07537 40.95878 45.35097 36.02482 73.57484
#> [5233] 47.38860 46.30574 42.89886 40.18501 51.08316 38.47707 42.00066 43.62494
#> [5241] 38.38922 44.65378 45.91256 41.86386 42.15635 39.77534 50.90808 46.36215
#> [5249] 55.74414 48.84789 39.09898 41.33538 46.61820 41.26850 53.60914 47.67704
#> [5257] 57.65869 48.02586 52.99742 60.00866 47.75043 62.08655 40.31769 47.23007
#> [5265] 65.66434 57.25619 61.11844 57.87879 66.77112 61.98004 73.19655 70.85959
#> [5273] 65.32637 69.97358 66.11595 68.74497 33.83822 38.67966 39.48467 42.86748
#> [5281] 30.22608 30.78180 41.06584 56.58093 31.54022 36.41742 27.36035 32.76724
#> [5289] 38.27589 38.27387 26.01509 38.01831 47.27670 35.61224 33.03681 44.26888
#> [5297] 44.88007 56.93611 76.02725 27.99844 41.47960 42.43965 45.89279 35.61224
#> [5305] 46.44475 52.91353 41.05234 48.16465 49.38377 33.27798 40.50761 22.54832
#> [5313] 42.81081 28.48744 41.80366       NA 33.72239       NA       NA 28.99927
#> [5321] 28.62974 32.14337 28.78043       NA 36.37541 37.30435 42.98272 34.94553
#> [5329] 41.51751 41.85069 43.57965 50.03196 36.25753 39.17696 56.35790 44.92155
#> [5337] 44.16504 46.08024 42.95952 45.73114 36.79968 31.81767 30.54060 56.63564
#> [5345] 53.07682 43.03986 41.16705 51.32409 34.89290 37.05126 32.47904 25.89705
#> [5353] 29.63559 35.46703 31.96837 38.56688 34.12745 41.08013 39.14798 33.35904
#> [5361] 31.64466 29.43172 28.03892 43.69015 37.63548 37.84920 59.45365 33.89608
#> [5369] 38.24347 53.44994 40.93741 39.22841 36.37361 43.61817 35.83746 34.42361
#> [5377] 29.26443 30.31035 30.75960 35.24722 42.21368 31.84647 36.04098 30.29172
#> [5385] 33.07685 32.28610 38.04370 51.34364 63.96590 34.65598 42.69687 61.11776
#> [5393] 32.79601 46.89100 41.26974 59.53991 29.68147 25.45720 39.06097 41.37391
#> [5401] 35.75719 49.73406 34.98605 57.37655 41.74015 31.19601 38.39412 50.41274
#> [5409] 32.75992 49.32235 36.60270 39.08299 37.06280 42.21828 38.39006 36.34659
#> [5417] 29.78970 32.45018 29.72109 62.36351 37.44664 32.95555 26.35644 29.02979
#> [5425] 37.10072 29.93274 29.77975 37.29286 33.49028 40.27709 38.33644 30.59810
#> [5433] 27.88157 35.99985 45.00883 40.09353 39.91159 32.15722 31.18178 30.58774
#> [5441] 29.84694 34.84705 42.64185 29.88981 46.67902 31.74372 44.72454 47.90561
#> [5449] 45.89279 30.93559 33.43752 36.27191 53.44524 56.51629 62.20490 59.70697
#> [5457] 66.50283 64.93629 64.21588 71.98078 53.70150 63.39077 63.00780 56.57905
#> [5465] 37.93604 45.80358 46.30574 52.01007 37.69870 37.20908 46.19626 61.72082
#> [5473] 47.23836 36.65399 34.15738 46.84293 51.84914 46.25077 31.84201 38.30293
#> [5481] 53.11895 40.82226 42.75361 50.52607 60.10528 59.75703 74.76370 34.58735
#> [5489] 50.81346 45.02685 47.79542 48.34731 56.13749 54.33554 53.18978 58.20959
#> [5497] 53.47142 48.26406 45.50684 30.48406 45.83173 34.33343 57.82751       NA
#> [5505] 39.44239       NA       NA 31.87527 35.35574 42.44086 36.21984       NA
#> [5513] 55.68384 53.74130 47.41653 41.65219 45.03748 39.60113 49.02864 45.73114
#> [5521] 54.00424 51.72100 60.10528 44.96467 46.52956 46.17680 46.42070 57.89667
#> [5529] 46.72722 42.81260 39.66788 56.99772 65.19206 58.95588 56.33254 62.38576
#> [5537] 45.44282 41.65219 39.79614 35.50595 38.53515 42.61080 35.70970 43.49477
#> [5545] 34.61777 41.58463 45.71903 41.47032 41.09727 35.51597 39.79437 44.62816
#> [5553] 42.17848 43.09693 77.43674 42.86829 43.37413 49.82859 46.64321 66.79737
#> [5561] 56.96877 49.81623 38.91569 37.58803 43.70984 38.50722 34.82549 38.94023
#> [5569] 46.83899 40.46228 42.27096 43.12544 36.28891 35.72822 45.71903 48.21721
#> [5577] 57.27348 43.44711 47.40768 69.05781 39.31770 50.01188 48.54097 75.74876
#> [5585] 35.86233 32.78877 59.22251 50.87937 40.95134 59.10823 43.03986 68.45709
#> [5593] 50.33333 42.52355 43.98694 59.51579 43.85301 52.81704 40.90063 48.23283
#> [5601] 50.37856 56.20546 47.04475 46.89478 32.74694 37.45754 42.06605 64.40008
#> [5609] 50.00491 35.44976 32.68900 39.00666 45.30975 37.08035 39.72910 50.51495
#> [5617] 36.90560 46.24901 47.77814 38.97384 32.44952 57.67731 52.44071 56.56726
#> [5625] 45.56301 46.37784 43.07938 38.56372 34.74295 40.70885 48.28790 46.57131
#> [5633] 50.89288 35.74732 51.28427 49.24752 51.46112 51.89654 41.79562 38.94477
#> [5641] 67.69384 71.20943 56.08833 56.21753 67.09038 69.17997 64.25724 71.78427
#> [5649] 54.00570 61.21394 71.04807 60.37430