Default correction parameters

Returns a list with the default parameterization for statistical correction.

defaultCorrectionParams()

Value

A list with the following items (default values in brackets):

methods: A list with the correction method for each variable (options are "unbias" (shift of the mean), "scaling" (factor multiplication), "quantmap" (empirical quantile mapping) or "none" (for no correction)). Defaults are:

MeanTemperature = "unbias"

MinTemperature = "quantmap"

MaxTemperature = "quantmap"

Precipitation = "quantmap"

MeanRelativeHumidity = "unbias"

Radiation = "unbias"

WindSpeed = "quantmap"

fill_wind [= TRUE]: A logical flag to fill wind speed values with uncorrected values when reference data is missing.

allow_saturated [= FALSE]: A logical flag to indicate whether relative humidity values above saturation (>100%) are permitted (bias correction is performed on specific humidity).

wind_height [= 10]: Wind measurement height (in m).

qstep [= 0.01]: a numeric value between 0 and 1. Quantile mapping is fitted only for the quantiles defined by quantile(0,1,probs=seq(0,1,by=qstep).

References

De Caceres M, Martin-StPaul N, Turco M, Cabon A, Granda V (2018) Estimating daily meteorological data and downscaling climate models over landscapes. Environmental Modelling and Software 108: 186-196.

See also

Author

Miquel De Cáceres Ainsa, CREAF