Default generation parameters

Returns a list with the default parameterization for weather generation.

defaultGenerationParams()

Value

A list with the following items (default values in brackets):

conditional [= "none"]: A string to indicate whether multi-year weather should be conditioned or not. If conditional = "arima", annual precipitation is conditioned on a stationary auto-regressive model. If conditional = "window", a moving-window is used to subset the years used to parametrize the weather generation algorithm for each target year (see parameter range_size_years). In this last case, annual precipitation is conditioned to a log-normal variate with parameters fitted from the selected years.

dry_wet_threshold [= 0.3]: Precipitation threshold (mm) for separating dry from wet days.

wet_extreme_quantile_threshold [= 0.8]: Quantile for separating wet from extremely wet days.

range_size_days [= 5]: Minimum half range size to select the subset of dates with DOY similar to the currently simulated.

range_size_years [= 12]: Half range size to select the subset of years in a moving-window around the current year (if conditional = "window").

n_knn_annual [= 100]: Number of years to be re-sampled using K-nearest neighbour for annual precipitation (if conditional = "arima" or conditional = "window").

adjust_annual_precip [= TRUE]: A logical flag to indicate that annual precipitation generated by the algorithm should be adjusted to fit either overall input annual precipitation or simulated annual precipitation.

min_ratio [= 0.9]: Minimum adjustment ratio for precipitation.

max_ratio [= 1.2]: Minimum adjustment ratio for precipitation.

References

Apipattanavis, S., G. Podesta, B. Rajagopalan, and R. W. Katz (2007), A semiparametric multivariate and multisite weather generator, Water Resour. Res., 43, W11401, doi:10.1029/2006WR005714.

Steinschneider S. & Brown C. (2013) A semiparametric multivariate, multisite weather generator with low-frequency variability for use in climate risk assessments. Water Resour. Res. 49, 7205-7220. doi:10.1002/wrcr.20528.

Author

Miquel De Cáceres Ainsa, CREAF

See also