Download data from Meteoclimatic network

DEPRECATED. Download data from the Spanish Automatic Stations Network (non-professional)

downloadMETEOCLIMATICstationlist(station_id = "ESCAT", spatial = TRUE)
downloadMETEOCLIMATICcurrentday(station_id = "ESCAT")

Arguments

station_id
A string vector containing station ids (the list of stations for which current day climatic data is available is given by downloadMETEOCLIMATICstationlist).

spatial
Returned object format. If TRUE (default), a SpatialPointsDataFrame-class object is returned. If FALSE, a data.frame is returned instead.

Details

Meteoclimatic is a non-professional automatic stations network, maintained by volunteers that share the data from their climatic stations. Data offered by these stations has not passed any quality control.

Value

Function downloadMETEOCLIMATICstationlist returns a SpatialPointsDataFrame-class object containing the list of Meteoclimatic weather stations for which data is available based on the station_id codes provided, or a data.frame with the same information if spatial = FALSE.

Function downloadMETEOCLIMATICcurrentday downloads recent weather (for the current day) from all currently available stations and returns a SpatialPointsDataFrame-class object with observations.

Note

Recommended package: meteospain (available at GitHub).

Only acummulated precipitation, maximum and minimum temperature and relative humidity are returned. The current implementation was inspired by function meteoclimaticR by Marc Lemus Cánovas.

References

Meteoclimatic should be acknowledged as author of information when using this data.

Author

Víctor Granda, EMF-CREAF

Miquel De Cáceres Ainsa, EMF-CREAF

See also