Download data from SMC

Download data from the Catalan automatic weather station network (XEMA from Servei Meteorològic de Catalunya)

downloadSMCvarmetadata(api, type = "current")
downloadSMCstationlist(api, date = NULL)
downloadSMCcurrentday(api, daily_meteoland = TRUE, variable_code=NULL,
                      station_id=NULL, date = Sys.Date(), verbose=TRUE)
downloadSMChistorical(api, dates, station_id=NULL, variable_code=NULL, export = FALSE, 
                      exportDir = getwd(), exportFormat = "meteoland/txt",
                      metadataFile = "MP.txt", verbose=TRUE)

Arguments

api
String with the SMC API key (the procedure to apply for an api key is explained in https://apidocs.meteocat.gencat.cat/ ).

type
Either 'current' or 'historical' for current day variable metadata or historical variable metadata, respectively.

daily_meteoland
When daily_meteoland = TRUE all the usual variables in meteoland are downloaded and aggregated at the daily scale. When daily_meteoland = FALSE, the user can choose the desired variables using variable_code.

variable_code
A character vector with variable codes to be queried. If NULL in downloadSMChistorical all the usual variables for meteoland are downloaded. Otherwise, the results are returned as a dara frame.

dates
An object of class Date with dates comprised within the last two years.

date
An object of class Date. By default the current day in the case of downloadSMCcurrentday(). In the case of downloadSMCstationlist() a date for which operational stations are queried.

station_id
A string vector containing station ids (the list of stations for which climatic series are available is given by downloadSMCstationlist()). If NULL, all available stations are queried. Otherwise, only the data corresponding to the specified stations will be returned.

export
If export = FALSE the downloaded data is stored in memory. Otherwise the result is written on the disk (using the format specified in exportFormat).

exportDir
Output directory for downloaded meteorology.

exportFormat
Format of meteorological data. Current accepted formats are "castanea" and "meteoland".

metadataFile
The name of the file that will store the meta data describing all written files.

verbose
Boolean flag to print process information.

Details

API key needs to be requested from SMC (https://apidocs.meteocat.gencat.cat/).

Value

Function downloadSMCstationlist returns a SpatialPointsDataFrame-class object containing the list of SMC operational weather stations for the date given.

Function downloadSMCvarmetadata returns a data frame with weather variables, their units and acronym to be used in queries (see parameter variable_code).

Function downloadSMCcurrentday downloads recent weather (the last 24h or the weather for a given date) from all currently available stations and returns data frame if daily_meteoland = FALSE or a SpatialPointsDataFrame-class object with observations aggregated at the daily scale otherwise.

Function downloadSMChistorical downloads historical daily weather corresponding to a given time period from a set (or all currently available) stations. Results are returned (or exported) after formatting data as a SpatialPointsMeteorology-class if variable_code = NULL, or as a data frame otherwise.

References

Servei Meteorològic de Catalunya (SMC) should be acknowledged as author of information when accessing weather data with these functions.

Author

Antoine Cabon, CTFC

Miquel De Cáceres Ainsa, CTFC

See also