Calibration and validation of interpolation procedures

Calibration and validation of interpolation procedures

interpolation_cross_validation(
  interpolator,
  stations = NULL,
  verbose = getOption("meteoland_verbosity", TRUE)
)

interpolator_calibration(
  interpolator,
  stations = NULL,
  update_interpolation_params = FALSE,
  variable = c("MinTemperature", "MaxTemperature", "DewTemperature", "Precipitation",
    "PrecipitationAmount", "PrecipitationEvent"),
  N_seq = seq(5, 30, by = 5),
  alpha_seq = seq(0.25, 10, by = 0.25),
  verbose = getOption("meteoland_verbosity", TRUE)
)

Arguments

interpolator

A meteoland interpolator object, as created by create_meteo_interpolator

stations

A vector with the stations (numeric for station indexes or character for stations id) to be used to calculate "MAE". All stations with data are included in the training set but predictive "MAE" are calculated for the stations subset indicated in stations param only. If NULL all stations are used in the predictive "MAE" calculation.

verbose

Logical indicating if the function must show messages and info. Default value checks "meteoland_verbosity" option and if not set, defaults to TRUE. It can be turned off for the function with FALSE, or session wide with options(meteoland_verbosity = FALSE)

update_interpolation_params

Logical indicating if the interpolator object must be updated with the calculated parameters. Default to FALSE

variable

A string indicating the meteorological variable for which interpolation parameters "N" and "alpha" will be calibrated. Accepted values are MinTemperature, MaxTemperature, DewTemperature, Precipitation (for precipitation with the same values for precipitation events an regression of precipitation amounts), PrecipitationAmount (for regression of precipitation amounts) and PrecipitationEvent (for precipitation events).

N_seq

Numeric vector with "N" values to be tested

alpha_seq

Numeric vector with "alpha"

Value

interpolation_cross_validation returns a list with the following items

errors: Data frame with each combination of station and date with observed variables, predicated variables and the total error (predicted - observed) calculated for each variable

station_stats: Data frame with error and bias statistics aggregated by station

dates_stats: Data frame with error and bias statistics aggregated by date

r2: correlation indexes between observed and predicted values for each meteorological variable

If update_interpolation_params is FALSE (default), interpolator_calibration returns a list with the following items

MAE: A numeric matrix with the mean absolute error values, averaged across stations, for each combination of parameters "N" and "alpha"

minMAE: Minimum MAE value

N: Value of parameter "N" corresponding to the minimum MAE

alpha: Value of parameter "alpha" corresponding the the minimum MAE

observed: matrix with observed values (meteorological measured values)

predicted: matrix with interpolated values for the optimum parameter combination

If update_interpolation_params is FALSE, interpolator_calibration

returns the interpolator provided with the parameters updated

Details

Function interpolator_calibration determines optimal interpolation parameters "N" and "alpha" for a given meteorological variable. Optimization is done by minimizing mean absolute error ("MAE") (Thornton et al. 1997). Function interpolation_cross_validation calculates average mean absolute errors ("MAE") for the prediction period of the interpolator object. In both calibration and cross validation procedures, predictions for each meteorological station are made using a leave-one-out procedure (i.e. after excluding the station from the predictive set).

Functions

interpolation_cross_validation():

Examples

# \donttest{
# example interpolator
data("meteoland_interpolator_example")

# As the cross validation for all stations can be time consuming, we are
# gonna use only for the first 5 stations of the 198
cv <- interpolation_cross_validation(meteoland_interpolator_example, stations = 1:5)
#> ℹ Starting Cross Validation process...
#> • Interpolating stations...
#> • Calculating R squared...
#> • calculating errors, MAE and bias for interpolated variables...
#> ✔ Cross validation done.

# Inspect the results
cv$errors
#> # A tibble: 150 × 21
#>    dates               station statio…¹ MinTe…² MaxTe…³ Range…⁴ Relat…⁵ Radiat…⁶
#>    <dttm>                <int> <chr>      <dbl>   <dbl>   <dbl>   <dbl>    <dbl>
#>  1 2022-04-01 00:00:00       1 C6        0.120   -0.628 -0.748   -1.43  -1.53   
#>  2 2022-04-02 00:00:00       1 C6       -0.296   -0.364 -0.0678  -2.12  -0.275  
#>  3 2022-04-03 00:00:00       1 C6        0.178   -0.919 -1.10    -1.97  -0.123  
#>  4 2022-04-04 00:00:00       1 C6       -0.164   -0.292 -0.129   -1.11  -0.168  
#>  5 2022-04-05 00:00:00       1 C6       -0.0108  -1.27  -1.26     0.323 -0.0116 
#>  6 2022-04-06 00:00:00       1 C6       -0.287   -0.690 -0.403   -1.45  -0.00222
#>  7 2022-04-07 00:00:00       1 C6       -0.773   -0.934 -0.161   -0.784  0.212  
#>  8 2022-04-08 00:00:00       1 C6        1.22    -0.710 -1.93    -1.77  -0.408  
#>  9 2022-04-09 00:00:00       1 C6       -0.599   -1.05  -0.451    0.349 -0.0391 
#> 10 2022-04-10 00:00:00       1 C6        1.08    -0.121 -1.20    -1.60  -0.597  
#> # … with 140 more rows, 13 more variables: Precipitation_error <dbl>,
#> #   MinTemperature_predicted <dbl>, MaxTemperature_predicted <dbl>,
#> #   RangeTemperature_predicted <dbl>, RelativeHumidity_predicted <dbl>,
#> #   Radiation_predicted <dbl>, Precipitation_predicted <dbl>,
#> #   MinTemperature_observed <dbl>, MaxTemperature_observed <dbl>,
#> #   RangeTemperature_observed <dbl>, RelativeHumidity_observed <dbl>,
#> #   Radiation_observed <dbl>, Precipitation_observed <dbl>, and abbreviated …
cv$station_stats
#> # A tibble: 5 × 20
#>   station stationID MinTempera…¹ MaxTe…² Range…³ Relat…⁴ Radia…⁵ MinTe…⁶ MaxTe…⁷
#>     <int> <chr>            <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1       1 C6             -0.0601 -0.489   -0.429  -1.03  -1.06     0.372   0.526
#> 2       2 C7             -1.37   -0.644    0.730   4.12  -0.338    1.40    0.671
#> 3       3 C8             -0.224   0.0301   0.254   0.108 -0.299    0.529   0.359
#> 4       4 C9             -0.257   0.519    0.776   2.81  -0.0291   0.701   0.696
#> 5       5 CC              0.178  -0.808   -0.986  -0.775 -1.93     0.530   0.895
#> # … with 11 more variables: RangeTemperature_station_mae <dbl>,
#> #   RelativeHumidity_station_mae <dbl>, Radiation_station_mae <dbl>,
#> #   TotalPrecipitation_station_observed <dbl>,
#> #   TotalPrecipitation_station_predicted <dbl>,
#> #   TotalPrecipitation_station_bias <dbl>,
#> #   TotalPrecipitation_station_relative_bias <dbl>,
#> #   DaysPrecipitation_station_bias <int>, …
cv$dates_stats
#> # A tibble: 30 × 19
#>    dates               MinTemp…¹ MaxTe…² Range…³ Relat…⁴ Radia…⁵ MinTe…⁶ MaxTe…⁷
#>    <dttm>                  <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 2022-04-01 00:00:00   -0.0883 -0.174  -0.0852   0.888 -1.25     0.305   0.578
#>  2 2022-04-02 00:00:00   -0.434  -0.226   0.207    1.86  -0.0601   0.434   0.791
#>  3 2022-04-03 00:00:00   -0.392  -0.477  -0.0851   0.436  0.129    0.819   0.600
#>  4 2022-04-04 00:00:00   -0.550  -0.197   0.354   -1.66   1.05     0.746   0.289
#>  5 2022-04-05 00:00:00   -0.172  -0.365  -0.193    1.43   0.454    0.548   0.768
#>  6 2022-04-06 00:00:00   -0.345   0.0511  0.396    1.53   0.304    0.527   0.475
#>  7 2022-04-07 00:00:00   -1.09   -0.188   0.899    2.04   0.639    1.24    0.455
#>  8 2022-04-08 00:00:00   -0.730  -0.236   0.493    1.98   0.262    1.40    0.488
#>  9 2022-04-09 00:00:00   -0.576  -0.422   0.154    0.148 -0.0608   0.833   0.654
#> 10 2022-04-10 00:00:00   -0.568   0.0402  0.609   -0.412 -1.41     1.29    0.831
#> # … with 20 more rows, 11 more variables: RangeTemperature_date_mae <dbl>,
#> #   RelativeHumidity_date_mae <dbl>, Radiation_date_mae <dbl>,
#> #   TotalPrecipitation_date_observed <dbl>,
#> #   TotalPrecipitation_date_predicted <dbl>,
#> #   TotalPrecipitation_date_bias <dbl>,
#> #   TotalPrecipitation_date_relative_bias <dbl>,
#> #   StationsPrecipitation_date_bias <int>, …
cv$r2
#> $MinTemperature
#> [1] 0.9720093
#> 
#> $MaxTemperature
#> [1] 0.9845019
#> 
#> $RangeTemperature
#> [1] 0.9557687
#> 
#> $RelativeHumidity
#> [1] 0.9591181
#> 
#> $Radiation
#> [1] 0.8943052
#> 
# }

# \donttest{
# example interpolator
data("meteoland_interpolator_example")

# As the calibration for all stations can be time consuming, we are gonna
# interpolate only for the first 5 stations of the 198 and only a handful
# of parameter combinations
calibration <- interpolator_calibration(
  meteoland_interpolator_example,
  stations = 1:5,
  variable = "MaxTemperature",
  N_seq = seq(10, 20, by = 5),
  alpha_seq = seq(8, 9, by = 0.25)
)
#> ℹ Total number of stations: 189
#> ℹ Number of stations with available data: 185
#> ℹ Number of stations used for MAE calculation: 5
#> ℹ Number of parameters combinations to test: 15
#> ℹ Starting evaluation of parameter combinations for "MaxTemperature"...
#> • Evaluating N: 10, alpha: 8...
#> • Evaluating N: 10, alpha: 8.25...
#> • Evaluating N: 10, alpha: 8.5...
#> • Evaluating N: 10, alpha: 8.75...
#> • Evaluating N: 10, alpha: 9...
#> • Evaluating N: 15, alpha: 8...
#> • Evaluating N: 15, alpha: 8.25...
#> • Evaluating N: 15, alpha: 8.5...
#> • Evaluating N: 15, alpha: 8.75...
#> • Evaluating N: 15, alpha: 9...
#> • Evaluating N: 20, alpha: 8...
#> • Evaluating N: 20, alpha: 8.25...
#> • Evaluating N: 20, alpha: 8.5...
#> • Evaluating N: 20, alpha: 8.75...
#> • Evaluating N: 20, alpha: 9...
#> ✔ Calibration done: Minimum MAE: 0.558727762077721; N: 10; alpha: 9

# we can update the interpolator params directly:
updated_interpolator <- interpolator_calibration(
  meteoland_interpolator_example,
  stations = 1:5,
  update_interpolation_params = TRUE,
  variable = "MaxTemperature",
  N_seq = seq(10, 20, by = 5),
  alpha_seq = seq(8, 9, by = 0.25)
)
#> ℹ Total number of stations: 189
#> ℹ Number of stations with available data: 185
#> ℹ Number of stations used for MAE calculation: 5
#> ℹ Number of parameters combinations to test: 15
#> ℹ Starting evaluation of parameter combinations for "MaxTemperature"...
#> • Evaluating N: 10, alpha: 8...
#> • Evaluating N: 10, alpha: 8.25...
#> • Evaluating N: 10, alpha: 8.5...
#> • Evaluating N: 10, alpha: 8.75...
#> • Evaluating N: 10, alpha: 9...
#> • Evaluating N: 15, alpha: 8...
#> • Evaluating N: 15, alpha: 8.25...
#> • Evaluating N: 15, alpha: 8.5...
#> • Evaluating N: 15, alpha: 8.75...
#> • Evaluating N: 15, alpha: 9...
#> • Evaluating N: 20, alpha: 8...
#> • Evaluating N: 20, alpha: 8.25...
#> • Evaluating N: 20, alpha: 8.5...
#> • Evaluating N: 20, alpha: 8.75...
#> • Evaluating N: 20, alpha: 9...
#> ✔ Calibration done: Minimum MAE: 0.558727762077721; N: 10; alpha: 9

# check the new interpolator have the parameters updated
get_interpolation_params(updated_interpolator)$N_MaxTemperature
#> [1] 10
get_interpolation_params(updated_interpolator)$alpha_MaxTemperature
#> [1] 9
# }