Writes grid meteorology to the disk

Functions to write grid meteorological data to the file system.

writemeteorologygrid(object, file, dates = NULL, format = "netCDF", 
                     byPixel = FALSE, chunksizes = NA, add = FALSE, overwrite = FALSE, 
                     verbose = FALSE)
writeemptymeteorologygrid(file, grid, proj4string, dates, 
                     byPixel = FALSE, chunksizes = NA, overwrite = FALSE, 
                     verbose = FALSE)
writemeteorologypixels(object, file, dates = NULL, format = "netCDF", 
                       byPixel = FALSE, chunksizes = NA, add = FALSE, overwrite = FALSE, 
                       verbose = FALSE)
writemeteorologygridpixel(file, index, data, verbose = FALSE)

Arguments

object

An object of class SpatialGridMeteorology-class or class SpatialPixelsMeteorology-class with the meteorological data to be written.

file

A string with the file name to be written.

dates

A Date vector object or a character string indicating the dates to be written.

format

Format of meteorological data. The only accepted format is "netCDF".

byPixel

Boolean flag to specify whether file will be written/read by pixels. This forces a different (suposedly) more efficient chunking based on time series data instead of daily grids.

chunksizes

Specifies the size of data chunks to be read/written. If set, this must be a vector of three integers corresponding to XYT.

add

Boolean flag to indicate that NetCDF exists and data should be added/replaced.

grid

An object of class GridTopology-class

proj4string

Object of class CRS.

overwrite

Boolean flag to force overwriting an existing NetCDF.

verbose

A logical flag to output process information in the console.

index

Integer indicating the grid index position to be written.

data

A data frame with meteorological data corresponding to a pixel.

Details

Functions writemeteorologygrid and writemeteorologypixels writes grided meteorological data (i.e. class SpatialGridMeteorology-class or class SpatialPixelsMeteorology-class, respectively) into a netCDF, with the possibility to overwrite existing data. Function writemeteorologygridpixel is meant to add/replace data in a netCDF corresponding to a specific pixel identified by its grid index. Function writemeteorologygrid creates an empty netCDF with the specified grid dimensions, coordinate reference system and dates.

Author

Miquel De Cáceres Ainsa, CREAF

See also