Download data from MeteoGalicia

DEPRECATED. Download data from the Galician Meterology Agency (MeteoGalicia)

downloadMGstationlist()
downloadMGhistorical(date_from, date_to, station_id = NULL,  
                    verbose = TRUE)
downloadMGcurrentday(station_id = NULL, daily = TRUE, verbose = TRUE)

Arguments

date_from, date_to

Strings or objects of class Date specifying first and last date of the desired period.

station_id

A string vector containing station ids (the list of stations presently operative is given by downloadMGstationlist). If NULL all stations with available data are returned.

verbose

Boolean flag to print process information.

daily

Boolean flag. Are data to be returned at a daily or hourly scale?

Details

See available data services of MeteoGalicia at https://www.meteogalicia.gal/web/RSS/rssIndex.action?request_locale=es.

Value

Function downloadMGstationlist returns a SpatialPointsDataFrame-class object containing the list of MeteoGalicia weather stations currently operative. Function downloadMGhistorical downloads data for the specified MG weather stations (or all) and dates and returns a SpatialPointsMeteorology-class object with the downloaded meteorology for each station (point). Function downloadMGcurrentday downloads recent weather data (the last 24h) from all currently available stations and returns data frame if daily = FALSE or a SpatialPointsDataFrame-class object with observations aggregated at the daily scale otherwise.

Note

Since ver. 1.0.1, weather data download functions included in meteoland make internal calls to functions in package meteospain. For an enhanced flexibility, users are recommended to call functions in meteospain themselves, and then to use function reshapemeteospain to generate data suitable for meteoland.

References

MeteoGalicia (from the Conselleria de Medio Ambiente, Territorio e Vivenda of Xunta de Galicia) should be acknowledged as source of information when using this data.

Author

Miquel De Cáceres Ainsa, CREAF

See also