Spatial and temporal coverage of interpolation data

Function interpolation.coverage calculates, for each meteorological variable, the number of stations with data per date or the number of dates with data per station in an object of class MeteorologyInterpolationData-class.

interpolation.coverage(object, type = "spatial", percent = FALSE)

Arguments

object
An object of class MeteorologyInterpolationData-class.

type
A string with the coverage summary to be produced (either "spatial" or "temporal").

percent
A boolean flag to indicate that percentages should be returned instead of counts.

Value

If type = "spatial" the function returns an object of class SpatialPointsDataFrame with the number (or percentage) of dates with data per station and meteorological variable. If type = "temporal" the function returns an object of class data.frame with the number (or percentage) of stations with data per day and meteorological variable.

References

De Caceres M, Martin-StPaul N, Turco M, Cabon A, Granda V (2018) Estimating daily meteorological data and downscaling climate models over landscapes. Environmental Modelling and Software 108: 186-196.

Author

Miquel De Cáceres Ainsa, CREAF

See also

Examples

data(exampleinterpolationdata)

#Number of days with data per station
head(interpolation.coverage(exampleinterpolationdata))

#> class       : SpatialPointsDataFrame 
#> features    : 6 
#> extent      : 347104, 382829, 4614634, 4659875  (xmin, xmax, ymin, ymax)
#> crs         : +proj=utm +zone=31 +datum=WGS84 +units=m +no_defs 
#> variables   : 7
#> names       : MinTemperature, MaxTemperature, Precipitation, RelativeHumidity, Radiation, WindSpeed, WindDirection 
#> min values  :           1288,           1288,          1288,                0,         0,         0,             0 
#> max values  :           1461,           1461,          1461,                0,         0,         0,             0 

#Number of stations with data per day
head(interpolation.coverage(exampleinterpolationdata, type = "temporal"))

#>            MinTemperature MaxTemperature Precipitation RelativeHumidity
#> 2000-01-01             13             13            14                1
#> 2000-01-02             13             13            14                1
#> 2000-01-03             13             13            14                1
#> 2000-01-04             13             13            14                1
#> 2000-01-05             13             13            14                1
#> 2000-01-06             13             13            14                1
#>            Radiation WindSpeed WindDirection
#> 2000-01-01         0         1             1
#> 2000-01-02         0         1             1
#> 2000-01-03         0         1             1
#> 2000-01-04         0         1             1
#> 2000-01-05         0         1             1
#> 2000-01-06         0         1             1