Potential evapotranspiration

Functions to calculate potential evapotranspiration using Penman or Penman-Monteith.

penman(
  latrad,
  elevation,
  slorad,
  asprad,
  J,
  Tmin,
  Tmax,
  RHmin,
  RHmax,
  R_s,
  u,
  z = 2,
  z0 = 0.001,
  alpha = 0.08,
  windfun = "1956"
)

penmanmonteith(rc, elevation, Tmin, Tmax, RHmin, RHmax, Rn, u = NA_real_)

Arguments

latrad

Latitude in radians.

elevation

Elevation (in m).

slorad

Slope (in radians).

asprad

Aspect (in radians from North).

J

Julian day, number of days since January 1, 4713 BCE at noon UTC.

Tmin

Minimum temperature (degrees Celsius).

Tmax

Maximum temperature (degrees Celsius).

RHmin

Minimum relative humidity (percent).

RHmax

Maximum relative humidity (percent).

R_s

Solar radiation (MJ/m2).

u

With wind speed (m/s).

z

Wind measuring height (m).

z0

Roughness height (m).

alpha

Albedo.

windfun

Wind speed function version, either "1948" or "1956".

rc

Canopy vapour flux (stomatal) resistance (s·m-1).

Rn

Daily net radiation (MJ·m-2·day-1).

Value

Potential evapotranspiration (in mm of water).

Details

The code was adapted from package `Evapotranspiration', which follows McMahon et al. (2013). If wind speed is not available, an alternative formulation for potential evapotranspiration is used as an approximation (Valiantzas 2006)

Functions

penmanmonteith(): Penman Monteith method

References

Penman, H. L. 1948. Natural evaporation from open water, bare soil and grass. Proceedings of the Royal Society of London. Series A. Mathematical and Physical Sciences, 193, 120-145.

Penman, H. L. 1956. Evaporation: An introductory survey. Netherlands Journal of Agricultural Science, 4, 9-29.

McMahon, T.A., Peel, M.C., Lowe, L., Srikanthan, R., McVicar, T.R. 2013. Estimating actual, potential, reference crop and pan evaporation using standard meteorological data: a pragmatic synthesis. Hydrology \& Earth System Sciences 17, 1331–1363. doi:10.5194/hess-17-1331-2013.

See also

Author

Miquel De Cáceres Ainsa, CREAF