Reads WindNinja results

Reads the wind fields generated by 'WindNinja' (http://www.firelab.org/project/windninja) for combinations of domain-level wind speed and wind direction classes.

readWindNinjaWindFields(
  filebase,
  resolution = "100m",
  directionClasses = c(0, 45, 90, 135, 180, 225, 270, 315),
  speedClasses = c(5, 15, 25),
  proj4string = CRS(as.character(NA))
)

Arguments

filebase

A string to indicate the template for accessing WindNinja files. Resolution, wind directions and wind speed class values are appended to this string to obtain the filename to read.

resolution

Resolution string.

directionClasses

A vector of wind speed directions (in degrees).

speedClasses

A vector of wind class values (in m/s).

proj4string

Object of class "CRS" with the projection string of wind field rasters.

Value

A list with the following items:

directionClasses: The vector of wind direction classes.

speedClasses: The vector of wind speed classes.

indexTable: A numeric matrix indicating the raster index of each combination of domain-level wind directions and wind speed classes.

windSpeed: An object of class SpatialGridDataFrame containing wind speed rasters (in m/s) for each combination of domain-level wind direction and wind speed.

windDirection: An object of class SpatialGridDataFrame containing wind direction rasters (in degrees from North) for each combination of domain-level wind direction and wind speed.

Note

WindNinja should be run with m/s as wind speed units and for all the combinations of domain-level wind speed and wind direction required.

References

Forthofer, J.M., Butler, B.W., Wagenbrenner, N.S., 2014. A comparison of three approaches for simulating fine-scale surface winds in support of wildland fire management. Part I. Model formulation and comparison against measurements. Int. J. Wildl. Fire 23, 969–981.

See also

Author

Miquel De Cáceres Ainsa, CREAF