Summaries of meteorological data

Summarizes the meteorology of a single location, a set of spatial points, pixels in a grid, or weather stations of interpolation data.

summarypoint(x, var, fun = mean, freq = NULL, dates = NULL, months = NULL, ...)
summarypoints(points, var, fun = mean, freq = NULL, dates = NULL, months = NULL, ...)
summarygrid(grid, var, fun = mean, freq = NULL, dates = NULL, months = NULL, ...)
summarypixels(pixels, var, fun = mean, freq = NULL, dates = NULL, months = NULL, ...)
summaryinterpolationdata(object, var, fun = mean, freq = NULL, dates = NULL, 
                         months = NULL, ...)

Arguments

x
A data frame with dates in rows and meteorological variables in columns.

points
An object of class SpatialPointsMeteorology-class with the coordinates and meteorological data of the locations for which summaries are desired. Alternatively, an object of class SpatialPointsDataFrame-class containing the meta data (columns dir, filename and possibly format) of meteorological files that will be sequentially read from the disk. Finally, points can also be a string pointing to a netCDF.

var
The name of the meteorological variable to be summarized.

fun
The function to be calculated on values of each point. If freq is specified, the function will be calculated by intervals.

freq
A string giving an interval specification (e.g., "week", "month", "quarter" or "year"). If NULL then no intervals are defined.

dates
An object of class Date to define the period to be summarized. If dates = NULL then all dates in points are processed.

months
A numeric vector to indicate the subset of months for which summary is desired (e.g. c(7,8) for July and August). This parameter allows studing particular seasons, when combined with freq. For example freq = "years" and months = 6:8 leads to summarizing summer months of each year.

...
Additional parameters to fun.

grid
An object of class SpatialGridMeteorology-class with the meteorological data for a grid, or a string pointing to a NetCDF.

pixels
An object of class SpatialPixelsMeteorology-class with the meteorological data for grid pixels, or a string pointing to a NetCDF.

object
An object of class MeteorologyInterpolationData-class.

Details

If var="ALL" then function summarypoints produces a summary of all variables with default statistics and returns an object of class SpatialPointsMeteorology.

Value

Function summarypoint returns a named vector of values with dates as names.

Functions summarypoints and summaryinterpolationdata return an object of class SpatialPointsDataFrame containing summaries (either one variable or several if freq is specified).

Functions summarygrid and summarypixels return an object of class SpatialGridDataFrame and SpatialPixelsDataFrame, respectively, containing the summaries analogously to summarypoints.

Author

Miquel De Cáceres Ainsa, CREAF

Antoine Cabon, CTFC

See also

Examples

data(examplegridtopography)
data(exampleinterpolationdata)

#Creates spatial topography points from the grid
p = 1:2
spt = as(examplegridtopography, "SpatialPointsTopography")[p]

#Interpolation of two points for the whole time period (2000-2003)
mp = interpolationpoints(exampleinterpolationdata, spt)

#> Processing point '1' (1/2) - done.
#> Processing point '2' (2/2) - done.

#PET sums by months
mp.sum = summarypoints(mp, var="PET", freq="months", fun=sum)

#>   Summarizing PET in 2 points...
#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%
#>   Arranging output...
#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

mp.sum

#> class       : SpatialPointsDataFrame 
#> features    : 2 
#> extent      : 360000, 360100, 4640900, 4640900  (xmin, xmax, ymin, ymax)
#> crs         : +proj=utm +zone=31 +ellps=WGS84 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs 
#> variables   : 48
#> names       : 2000-01-01, 2000-02-01,       2000-03-01,       2000-04-01,       2000-05-01, 2000-06-01, 2000-07-01, 2000-08-01, 2000-09-01, 2000-10-01,       2000-11-01,       2000-12-01, 2001-01-01,       2001-02-01,       2001-03-01, ... 
#> min values  :         NA,         NA, 78.9162423918269, 100.669725972632, 146.590523744577,         NA,         NA,         NA,         NA,         NA, 27.5296952566443, 18.4519158441549,         NA, 37.5410031751692, 79.8200553264094, ... 
#> max values  :         NA,         NA, 81.4547739646298, 101.922397207274, 147.028249622689,         NA,         NA,         NA,         NA,         NA, 29.8513681893117, 20.6740404840618,         NA, 39.9994829528526, 82.1779613810003, ...